% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEst.R
\name{allEst}
\alias{allEst}
\title{Estimates theta with multiple specifications}
\usage{
allEst(catObjs = list(), resp)
}
\arguments{
\item{catObjs}{A list of \code{Cat} objects of the same class (i.e., LTM, TPM, GRM, GPCM) specifying different estimation routines}

\item{respondents}{A matrix of response profiles}
}
\value{
The function \code{allEst} returns a vector.  Each \code{Cat} object returns a column and each respondent returns a row.


This function is to allow users to look at data computed before \code{Cat} technologies were availabe and and analyze multpile types of analyses on mulptile resondents to achieve a theta estimate.
}
\description{
Takes response profiles from multiple respondents and multiple Cat objects and returns a set of theta estimates
}
\details{
The function takes a \code{Cat} object and generates an estimation for \code{theta} using \code{MAP}, \code{EAP}, etc. 
The user must store all \code{Cat} objects in a single list.
}
\examples{

 ## Loading ltm Cat object
data(grm_cat)
   
## load sample data
 
respondents<-apply(thetaValue[1:3], 2, simulateRespondents, cat=grm_cat, n=10)

#Choose your \\code{cat} object and the types of estimation you would like to use
 grm_MAP<-grm_cat
 grm_MAP@estimation<-"MAP"
 grm_EAP<-grm_cat
 grm_EAP@estimation<-"EAP"


grmList<-list(grm_MAP, grm_EAP)
## Create a threshold for the estimation

## You must set some sort of CheckStopRule
grm_cat@lengthThreshold<-4

## Run 

allEst(catObjs=grmList, resp=respondents)

}
\seealso{
\code{\link{Cat-class}}, \code{\link{apply}}, \code{\link{selectItem}}, \code{\link{store}}
}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery, Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil, Jaerin Kim, Dominique Lockett
}
