% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFish.R
\name{allFish}
\alias{allFish}
\title{A function that finds total information that can be gained via answered items .}
\usage{
allFish(catObjs = list(), theta, resp)
}
\arguments{
\item{catObjs}{A list of \code{Cat} objects of the same class.}

\item{theta}{A vector of numerics representing the true value of theta.}

\item{resp}{One or many full response profiles.}
}
\value{
The function \code{allFish} returns a dataframe.  Each \code{Cat} object returns a column and each respondent returns a row.
}
\description{
Takes in a a \code{Cat} object, a set of respondents, and their corresponding \code{theta} values, to calculate the amount of information which can be extracted given an adaptive battery
}
\details{
The function takes a \code{Cat} object, \code{theta}, and response profiles. 
The user defines the selection type, estimation type, etc. so that the questions can be applied adaptively
These adaptive profiles are then used to calculate the total inforamtion gained for a respondent for all answered
items, conditioned on \code{theta}.
}
\examples{

 ## Loading ltm Cat object
data(grm_cat)
   
## load sample data
 
respondents<-apply(thetaValue[1:3], 2, simulateRespondents, cat=grm_cat, n=10)

#Choose your \\code{cat} object and the types of estimation you would like to use
 grm_MAP<-grm_cat
 grm_MAP@estimation<-"MAP"
 grm_EAP<-grm_cat
 grm_EAP@estimation<-"EAP"


grmList<-list(grm_MAP, grm_EAP)
## Create a threshold for the estimation

## You must set some sort of CheckStopRule
grm_cat@lengthThreshold<-4

## Run 

allFish(catObjs=grmList, resp=respondents)
}
\seealso{
\code{\link{Cat-class}}, \code{\link{fisherTestInfo}}, \code{\link{selectItem}}, \code{\link{store}}
}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery, Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil, Jaerin Kim, Dominique Lockett
}
