% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimateSE}
\alias{estimateSE}
\title{Standard error of the respondent's ability parameter estimate}
\usage{
estimateSE(catObj)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}
}
\value{
The function returns a numeric for the standard error for \eqn{\theta_j}
}
\description{
Estimates the standard error for the ability parameter estimate
}
\details{
The EAP estimator:
  The MAP estimator (This is implemented only for the normal prior.)
  The MLE  estimator (When MLE can't be calculated... estimationDefault)
  The WLE estimator (Brent method)
}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled C++ code.
}
\examples{
\dontrun{
## Prior calculation using Cat object of the ltm model
## specifying different distributions
data(npi)
cat <- ltmCat(npi)
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{estimateTheta}} for estimation of \eqn{\theta}
}

