% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimateTheta}
\alias{estimateTheta}
\title{Estimate of the ability parameter}
\usage{
estimateTheta(cat_df)
}
\arguments{
\item{cat}{An object of \code{Cat} class}
}
\value{
A vector consisting of the expected value of the ability parameter
}
\description{
This function takes a Cat object and returns the expected value of the ability parameter conditioned on the observed answers and the item calibrations.
}
\details{
The expected value of \eqn{\theta_j} is:
\eqn{\frac{\int_{-\infty}^\infty \theta_j \eqn{L}(\theta_j) \pi(\theta_j) d\theta_j}{\int_{-\infty}^\infty \eqn{L}(\theta_j) \pi(\theta_j) d\theta_j}},
where \eqn{L}(\theta_j) is the likelihood function and \pi(\theta_j) is the prior distribution for \eqn{\theta_j}


 Note: When MLE would not work, estimateTheta is calculated according to the estimationDefault slot in Cat object.

 Note: This method will only be available using the normal prior distribution
}

