% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimateThetaMAP.R
\name{estimateThetaMAP}
\alias{estimateThetaMAP}
\title{Computerized Adaptive Testing Survey Maximum A Posteriori Function}
\usage{
estimateThetaMAP(cat)
}
\arguments{
\item{cat}{an object of \code{CATsurv} class.}
}
\value{
A scalar valur representing the maximum value of the posterior density of theta.
}
\description{
This function estimates the maximum value of the posterior density of theta.
}
\author{
Josh W. Cutler and Jacob M. Montgomery
}
\seealso{
\code{\link{likelihood}},\code{\link{prior}}, \code{\link{estimateTheta}}, \code{\link{estimateSE}}, \code{\link{expectedPV}}, \code{\link{nextItem}}, \code{\link{storeAnswer}}, \code{\link{debugNextItem}}
}

