% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_qualtrics_results}
\alias{ex_qualtrics_results}
\title{Example Qualtrics Data for Adaptive Inventory}
\format{An object of class \code{data.frame} with 6 rows and 9 columns.}
\usage{
data(ex_qualtrics_results)
}
\description{
Example data when including an adaptive inventory on a Qualtrics survey
}
\details{
This data set includes four example responses to a length four adaptive inventory
 for the agreeableness battery.

See \code{\link{agree_cat}} for details regarding the Cat object structure and battery items.

See \code{\link{readQualtrics}} for how to clean adaptive inventory response profiles.
}
\seealso{
\code{\link{agree_cat}}, \code{\link{readQualtrics}}
}
\keyword{datasets}
