% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_qualtrics_results_multiple}
\alias{ex_qualtrics_results_multiple}
\title{Example Qualtrics Data for Multiple Adaptive Inventories}
\format{An object of class \code{data.frame} with 6 rows and 17 columns.}
\usage{
data(ex_qualtrics_results_multiple)
}
\description{
Example data when including multiple adaptive inventory on a Qualtrics survey
}
\details{
This data set includes four example responses to length four adaptive inventories
 for both the agreeableness and neuroticism batteries.

See \code{\link{agree_cat}} and \code{\link{neuro_cat}} for details regarding the Cat object structure and battery items.

See \code{\link{readQualtrics}} for how to clean adaptive inventory response profiles.
}
\seealso{
\code{\link{agree_cat}}, \code{\link{neuro_cat}},\code{\link{readQualtrics}}
}
\keyword{datasets}
