% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expectedInfo.R
\name{expectedInfo}
\alias{expectedInfo}
\title{Computerized Adaptive Testing Survey Expected Information Function}
\usage{
expectedInfo(cat, theta, items)
}
\arguments{
\item{cat}{an object of \code{CATsurv} class.}

\item{theta}{vector consisting of each respondent's position on the latent scale of interest.}

\item{item-level}{parameters data frame containing discrimination parameter, guessing parameter, difficulty parameter, and answer for each item.}
}
\value{
A scalar value representing the expected Fisher's information.
}
\description{
This function calculates the expected Fisher's information for the given theta estimate over these \eqn{k-1} items.
}
\details{
The scalar value of expected Fisher's information calculated by van der Linden and Parsley sum formula where nominator is calculated by \deqn{(\frac{\partial p(\theta)}{\partial \theta})^2=Da_i(1-c_i)\frac{exp[Da_i(\theta-b_i)]}{(1+exp[Da_i(\theta-b_i)])^2}} and denominator is calculated by calling the three.pl function for \eqn{p_i(\theta_j)} and \eqn{q_i(\theta_j)}.
}
\author{
Josh W. Cutler and Jacob M. Montgomery
}
\seealso{
\code{\link{likelihood}},\code{\link{prior}}, \code{\link{estimateTheta}}, \code{\link{estimateSE}}, \code{\link{expectedPV}}, \code{\link{nextItem}}, \code{\link{storeAnswer}}, \code{\link{debugNextItem}}
}

