% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expectedKL}
\alias{expectedKL}
\title{Expected Kullback-Leibeler information}
\usage{
expectedKL(catObj, item)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}

\item{item}{An integer indicating the index of the question item}
}
\value{
The function returns a numeric indicating the KL information for the desired item, given the current answer profile and ability parameter estimate
}
\description{
Calculates the expected Kullback-Leibeler information for an individual question item
}
\details{
Binary details (Due to the conditional independence assumption, we only need to calculate the expected value for potential new items.)
}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled C++ code.
}
\examples{
\dontrun{
## Prior calculation using Cat object of the ltm model
## specifying different distributions
data(npi)
cat <- ltmCat(npi)
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{likelihoodKL}} and/or \code{\link{posteriorKL}} for alternative KL methods
}

