% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expectedObsInf}
\alias{expectedObsInf}
\title{Expected Observed Information}
\usage{
expectedObsInf(catObj, item)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}

\item{item}{An integer indicating the index of the question item}
}
\value{
The function returns a numeric value of the expected information
}
\description{
Calculates the expected information, which is the observed information attained from a specific response set times the probability of that profile occurring
}
\details{

}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled C++ code.
}
\examples{
\dontrun{
## Prior calculation using Cat object of the ltm model
## specifying different distributions
data(npi)
cat <- ltmCat(npi)
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{estimateTheta}} for calculation of \eqn{\theta} and 
  \code{\link{obsInf}} for observed information calculation
}

