% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expectedPV}
\alias{expectedPV}
\title{Expected Posterior Variance}
\usage{
expectedPV(catObj, item)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}

\item{item}{An integer indicating the index of the question item}
}
\value{
A numeric value indicating a respondent's expected posterior variance for a yet to be asked question item
}
\description{
Estimates the expected posterior variance for a respondent's estimated ability parameter for an item yet to be answered based on a respondent's ability parameter estimate from the already-answered items
}
\details{

}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled C++ code.
}
\examples{
\dontrun{
## Prior calculation using Cat object of the ltm model
## specifying different distributions
data(npi)
cat <- ltmCat(npi)
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{probability}} for probability of responses to individual question item

\code{\link{estimateTheta}} for estimation of \eqn{\theta}
}

