% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fisherTestInfo}
\alias{fisherTestInfo}
\title{Fisher's Test Information}
\usage{
fisherTestInfo(catObj)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}
}
\value{
The total information gained for a respondent, given a specific answer set and a value of \eqn{theta}.
}
\description{
Calculates the total information gained for a respondent \eqn{j} for all answered items, conditioned on \eqn{theta}.
}
\details{

}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled C++ code.
}
\examples{
\dontrun{
## Prior calculation using Cat object of the ltm model
## specifying different distributions
data(npi)
cat <- ltmCat(npi)
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{fisherInf}} for calculation of Fisher's information for an individual question item
}

