% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromJSONCat.R
\name{fromJSONCat}
\alias{fromJSONCat}
\alias{fromJSONCat,json-method}
\alias{fromJSONCat,character-method}
\title{Convert JSON object to Cat object}
\usage{
\S4method{fromJSONCat}{json}(jsonObj)

\S4method{fromJSONCat}{character}(jsonObj)
}
\arguments{
\item{jsonObj}{A \code{json} object with elements matching Cat object slots}
}
\value{
The function \code{fromJSONCat} returns an object of class \code{Cat} with slots populated with the corresponding information stored in the \code{jsonObj}.

See \code{\link{Cat-class}} for required \code{Cat} object slots.
}
\description{
This function  object of class \code{Cat}.
}
\details{
The \code{jsonObj} argument of the function \code{fromJSONCat} is an object of class \code{json} from the \code{rjson} package.  This function simply populates a \code{Cat} object with the information provided in the \code{jsonObj}, thus element and slot names must match.
}
\seealso{
\code{\link{Cat-class}}
}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery, Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
