% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fticPlots.R
\name{ftic}
\alias{ftic}
\title{Plotting FisherTestInfo Curves (FTIC)}
\usage{
ftic(object, theta_range = seq(-3, 3, 0.1), ...)
}
\arguments{
\item{object}{A \code{Cat} object}

\item{theta_range}{A numerical vector of thetas at which the probabilities are to be calculated 

 @return A list with components,
\itemize{
\item The \code{Cat} object from which the fisherInf values were calculated
\item \code{fticInfo} A matrix of theta values and their corresponding fisherTestInfo values
}}
}
\description{
This function plots the Fisher test information curves for an object of class \code{Cat}.
}

