% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cat-class.R
\name{getCatSurv}
\alias{getCatSurv}
\title{Access slots within \cod{Cat} class objects}
\usage{
getCatSurv(object = "Cat")
}
\arguments{
\item{object}{An object of class \code{Cat} whose slot we wish to access}
}
\value{
The value of the specified slot of the user-provided \code{Cat}
}
\description{
These functions allow us to access to the values currently stored in a given slot of an object of class \code{Cat} with the user-provided values.
}
\note{
all
}
\author{
Josh W. Cutler: \email{josh@zistle.com} and Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
\seealso{
\code{\link{Cat}}
}

