% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grm_cat}
\alias{grm_cat}
\title{grm Cat Object}
\format{
An object of class \code{Cat}.  See \code{\link{Cat-class}} for more details.
}
\usage{
data(grm_cat)
}
\description{
An object of class \code{Cat} created using the \code{grmCat} function with the \code{nfc} dataset.
}
\examples{
\dontrun{
## How this Cat object was created
data(nfc)
grm_cat <- grmCat(nfc, quadraturePoints = 100)
}

## How to load this Cat object for usage
data(grm_cat)

}
\seealso{
\code{\link{Cat-class}}, \code{\link{grmCat}}, \code{\link{nfc}}
}
\keyword{datasets}
