% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iicPlots.R
\name{iic}
\alias{iic}
\title{Plotting Item Information Curves (IIC)}
\usage{
iic(object, theta_range = seq(-3, 3, 0.1), question, ...)
}
\arguments{
\item{object}{A \code{Cat} object}

\item{theta_range}{A numerical vector of thetas at which the probabilities are to be calculated}

\item{question}{A single numeric indicating the question/item for which the curves are to be drawn.

 @return A list with components,
\itemize{
\item The \code{Cat} object from which the fisherInf values were calculated
\item \code{question} A single numeric indicating the question/item from which the curves were drawn
\item \code{iicInfo} A matrix of theta values and their corresponding fisherInf values
}}
}
\description{
This function plots the item information curves for an object of class \code{Cat}.
}

