% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonCat.R
\docType{methods}
\name{jsonCat}
\alias{jsonCat}
\alias{jsonCat,json-method}
\title{Convert JSON object to Cat object}
\usage{
\S4method{jsonCat}{json}(jsonObj)
}
\arguments{
\item{jsonObj}{A \code{json} object with elements matching Cat object slots}
}
\value{
The function \code{jsonCat} returns an object of class \code{Cat} with slots populated with the corresponding information stored in the \code{jsonObj}.

See \code{\link{Cat-class}} for required \code{Cat} object slots.
}
\description{
This function  object of class \code{Cat}.
}
\details{
The \code{jsonObj} argument of the function \code{jsonCat} is an object of class \code{json} from the \code{rjson} package.  This function simply populates a \code{Cat} object with the information provided in the \code{jsonObj}, thus element and slot names must match.
}
\examples{
## Creating json object
library(jsonlite)
json_cat <- toJSON(list(guessing = rep(0, 10),
discrimination = rnorm(n = 10, 0, 1.5),
difficulty = rnorm(n = 10, 0, 3),
answers = sample(c(0,1), size = 10, replace = TRUE),
priorName = "NORMAL",
priorParams = c(0,1),
lowerBound = -5,
upperBound = 5,
model = "ltm",
estimation = "EAP",
estimationDefault = "EAP",
selection = "EPV",
z = .9,
lengthThreshold = NA,
seThreshold = NA,
infoThreshold = NA,
gainThreshold = NA,
lengthOverride = NA,
gainOverride = NA))

cat <- jsonCat(json_cat)


}
\seealso{
\code{\link{Cat-class}}
}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery, Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
