% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihoodKL}
\alias{likelihoodKL}
\title{Expected Kullback-Leibeler information, weighted by the likelihood}
\usage{
likelihoodKL(catObj, item)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}

\item{item}{An integer indicating the index of the question item}
}
\value{
A value indicating the LKL information for the desired item, given the current answer profile and ability estimate.
}
\description{
Calculate the expected Kullback-Leibeler information, weighted by the likelihood
}
\details{
The LKL calculation follows the same procedure as \code{expectedKL}, except it requires weighting the different potential values of \eqn{\theta_0} by the likelihood.
 Thus, the equation is

 Binary details:
 
 Due to the conditional independence assumption, we only need to calculate the expected value for potential new items.
}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled C++ code.
}
\examples{
\dontrun{
## Prior calculation using Cat object of the ltm model
## specifying different distributions
data(npi)
cat <- ltmCat(npi)
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{expectedKL}} and/or \code{\link{posteriorKL}} for alternative KL methods
}

