% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lookAhead}
\alias{lookAhead}
\title{Look Ahead to Select Next Item}
\usage{
lookAhead(catObj, item)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}

\item{item}{A numeric indicating the question item the respondent is currently answering.}
}
\value{
A vector of values indicating the possible subsequent questions
}
\description{
Selects the next item that would be asked for all possible response options to the question the respondent is currently answering
}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled C++ code.
}
\examples{
\dontrun{
## Prior calculation using Cat object of the ltm model
## specifying different distributions
data(npi)
cat <- ltmCat(npi)
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{selectItem}} for selection method information
}

