% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ltm_cat}
\alias{ltm_cat}
\title{ltm Cat Object}
\format{An object of class \code{Cat}.  See \code{\link{Cat-class}} for more details.}
\usage{
data(ltm_cat)
}
\description{
An object of class \code{Cat} created using the \code{ltmCat} function with the \code{npi} dataset.
}
\examples{
\dontrun{
## How this Cat object was created
data(npi)
ltm_cat <- ltmCat(npi, quadraturePoints = 100)
}

## How to load this Cat object for usage
data(ltm_cat)

}
\seealso{
\code{\link{Cat-class}}, \code{\link{ltmCat}}, \code{\link{npi}}
}
\keyword{datasets}
