% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmCat.R
\name{ltm-class}
\alias{ltm-class}
\title{Computerized Adaptive Testing Latent Trait Model}
\arguments{
\item{data}{A \code{data.frame} of manifest variables or an object of class \code{ltm}.}

\item{quadraturePoints}{A numeric to be passed into the \code{ltm} function indicating the number of Gauss-Hermite quadrature points.  Only applicable when \code{data} is a \code{data.frame}.  Default value is \code{21}.}

\item{...}{arguments to be passed to methods. For more details about the arguments, see \code{ltm} in the \code{ltm} package.}
}
\value{
The function \code{ltmCat} returns an object of class \code{Cat} with changes to the following slots:
\itemize{
\item \code{difficulty} A vector consisting of difficulty parameters for each item.
\item \code{discrimination} A vector consisting of disrimination parameters for each item.
\item \code{model} The string \code{"ltm"}, indicating this \code{Cat} object corresponds to a latent trait model.
}
}
\description{
This function fits the latent trait model for binary data and populates the fitted values for discimination and difficulty parameters to an object of class \code{Cat}.
}
\details{
The \code{data} argument of the function \code{ltmCat} is either a \code{data.frame} or an object of class \code{ltm} from the \code{ltm} package.  If it is a \code{data.frame} each row represents a respondent and each column represents a question item.  If it is an object of the class \code{ltm}, it is output from the \code{ltm} function in the \code{ltm} package.

The \code{quadraturePoints} argument of the function \code{ltmCat} is used only when the \code{data} argument is of class \code{data.frame}.  \code{quadraturePoints} is then passed to the \code{ltm} function from the \code{ltm} package when fitting the latent trait model to the data and is used when approximating the value of integrals.
}
\note{
In case the Hessian matrix at convergence is not positive definite try to use \code{start.val = "random"}.
}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery, Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{Cat}} for information on all \code{Cat} slots and their default values

\code{\link{tpmCat}} for an alternative model fit to binary data
}

