% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nfa_cat}
\alias{nfa_cat}
\title{Need for Affect Cat Object}
\format{An object of class \code{Cat} of length 1.}
\source{
See \url{https://dataverse.harvard.edu/dataverse/pdsl} for the raw YouGov and/or Qualtrics data.

See \url{https://wc.wustl.edu/american-panel-survey} for raw TAPS data.
}
\usage{
data(nfa_cat)
}
\description{
Cat model created for the 26 item Need for Affect inventory
}
\references{
Maio, Gregory R., and Victoria M. Esses. "The need for affect: Individual
 differences in the motivation to approach or avoid emotions." 
 Journal of personality 69.4 (2001): 583-614.
}
\keyword{datasets}
