\name{nfc}
\alias{nfc}
\docType{data}
\title{
%%   ~~ Need For Cognition data from TAPS~~
}
\description{
%%  ~~ Respondents' answers to 18 NFC questions (reduced version of NFC)  ~~
}
\usage{data("nfc")}
\format{
  A data frame with 4043 observations on the following 20 variables.
  \describe{
    \item{\code{X}}{a numeric vector}
    \item{\code{NFC1}}{a numeric vector}
    \item{\code{NFC4}}{a numeric vector}
    \item{\code{NFC10}}{a numeric vector}
    \item{\code{NFC12}}{a numeric vector}
    \item{\code{NFC15}}{a numeric vector}
    \item{\code{NFC16}}{a numeric vector}
    \item{\code{NFC19}}{a numeric vector}
    \item{\code{NFC21}}{a numeric vector}
    \item{\code{NFC22}}{a numeric vector}
    \item{\code{NFC23}}{a numeric vector}
    \item{\code{NFC24}}{a numeric vector}
    \item{\code{NFC29}}{a numeric vector}
    \item{\code{NFC31}}{a numeric vector}
    \item{\code{NFC32}}{a numeric vector}
    \item{\code{NFC33}}{a numeric vector}
    \item{\code{NFC39}}{a numeric vector}
    \item{\code{NFC40}}{a numeric vector}
    \item{\code{NFC43}}{a numeric vector}
    \item{\code{sample}}{a numeric vector}
  }
}
\details{
%%  ~~ Questions were asked in the fall of 2011 ~~
}
\source{
%%  ~~ TAPS is conducted by Washington University in St.Louis ~~
}
\references{
%%  ~~ Jacob Montgomery ~~
}
\examples{
data(nfc)
## maybe str(nfc) ; plot(nfc) ...
}
\keyword{datasets}
