% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nte}
\alias{nte}
\title{Need to Evaluate}
\format{
An object of class \code{data.frame} with 4005 rows and 16 columns.
}
\usage{
data(nte)
}
\description{
4005 full response profiles to 16 Need to Evaluate inventory.
Observations come from one of three administrations of the survey: The American Panel Survey, 
Amazon's Mechanical Turk  in fall of 2014, or Amazon's Mechanical Turk in spring of 2015.
}
\details{
See \code{\link{nte_cat}} for detailed information on question wordings and response options.
}
\references{
Jarvis, W. Blair G., and Richard E. Petty.
"The need to evaluate." Journal of personality and social psychology 70.1 (1996): 172.
}
\seealso{
\code{\link{nte_cat}}
}
\keyword{datasets}
