% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/observedInfo.R
\name{observedInfo}
\alias{observedInfo}
\title{Computerized Adaptive Testing Survey Observed Information Function}
\usage{
observedInfo(cat, theta, items)
}
\arguments{
\item{cat}{an object of \code{CATsurv} class.}

\item{theta}{vector consisting of each respondent's position on the latent scale of interest.}

\item{item-level}{parameters data frame containing discrimination parameter, guessing parameter, difficulty parameter, and answer for each item.}
}
\value{
A scalar value representing the expected Fisher's information.
}
\description{
This function calculates the observed information for the given theta estimate over these \eqn{k-1} items.
}
\details{
The scalar value of observed information measure is the negative of second derivative of the log likelihood function.
}
\author{
Josh W. Cutler and Jacob M. Montgomery
}
\seealso{
\code{\link{expectedInfo}}
}

