% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obsInf}
\alias{obsInf}
\title{Observed Information}
\usage{
obsInf(cat_df, theta, item)
}
\arguments{
\item{cat_df}{An object of class \code{Cat}}

\item{theta}{A double indicating the value of \eqn{\theta_j}}

\item{item}{An integer indicating the index of the question}
}
\value{
The value of the observed information of the likelihood, given \eqn{\theta}, for a specific question.
}
\description{
This function calculates the observed information of the likelihood evaluated at the input value \eqn{\theta} for a specific item.
}
\details{
The observed information is equivalent to the negative second derivative of the log-likelihood.
  Note: This function should never be called when the respondent has answered no questions.

  For the binary case:

  \deqn{\lambda_{\theta \theta} = b_i^2\Big(\frac{P_{ij} - c_i}{1-c_i} \Big)^2\frac{Q_{ij}}{P_{ij}}}{\lambda_{\theta \theta} = b_i^2 ((P_ij - c_i)/(1 - c_i))^2 (Q_ij)/(P_ij)}

  For the categorical case:

  \deqn{P_{ijk} = P^*_{ijk} - P^*_{ij,k-1}}{P_{ijk} = P*_{ijk} - P*_{ij,k-1}} and \deqn{P^*_{ijk} = Pr(y_{ij}<k|\theta_j)}{P*_ijk = Pr(y_ij < k | \theta_j)} as defined in \code{Likelihood}. 
  \deqn{Q_{ijk}^*= 1- P^*_{ijk}}{Q*_ijk = 1 - P*_ijk}, \deqn{w_{i,k-1} = P^*_{i,k-1}Q^*_{i,k-1}}{w_{i,k-1} = P*_{i,k-1} Q*_{i,k-1}}, and \deqn{w_{ik}=P^*_{ik}Q^*_{ik}}{w_ik = P*_ik Q*_ik} as defined in \code{dLL}.

  To calculate the categorical observed information requires evaluating the below equation for question \eqn{i}, assuming response \eqn{k}:  

  \deqn{\frac{\partial^2L}{\partial\theta_j^2} & -\beta_i^2 \Big [ \frac{-w_{i,k-1}(Q^*_{i,k-1}-P^*_{i,k-1}) + w_{ik}(Q^*_{ik}-P^*_{ik})}{P_{ik}}-\frac{(w_{ik}-w_{i,k-1})^2}{P_{ik}^2} \Big ]}{(\partial^2L)/(\partial\theta_j^2) = -\beta_i^2 [(-w_{i,k-1}(Q*_{i,k-1} - P*_{i,k-1}) + w_{ik}(Q*_{ik} - P*_{ik}))/(P_{ik}) - ((w_{ik} - w_{i,k-1})^2)/(P_{ik}^2)]}
}
\examples{

## binary (ltm)

 data("npi")
 ltm_data <- npi[1:100, ]
 ltm_cat <- ltmCat(ltm_data, quadraturePoints = 100)
 ltm_cat@answers <- as.numeric(ltm_data[1,])
 
 obsInf(ltm_cat, 1, 1)
 
## binary (tpm)

 data("AMTknowledge")
 tpm_data <- AMTknowledge[1:100, ]
 tpm_cat <- tpmCat(tpm_data, quadraturePoints = 100)
 tpm_cat@answers <- as.numeric(tpm_data[1,])
 
 obsInf(tpm_cat, 1, 1)
 
## categorical (grm) 

 data("nfc")
 poly_data <- nfc[1:100, ]
 poly_cat <- grmCat(poly_data, quadraturePoints = 100)
 poly_cat@answers <- as.numeric(poly_data[1,])
 
 obsInf(poly_cat, 1, 1)
 
}
\seealso{
\code{\link{probability}} and/or \code{\link{likelihood}} for calculation of P and P*;  
  \code{\link{estimateTheta}} for calculation of \eqn{\theta};  
  \code{\link{expectedObsInf}} for further application of observed information
}

