% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\docType{methods}
\name{oracle}
\alias{oracle}
\alias{oracle,Cat-method}
\title{Find Answer Profile that Minimizes Bias}
\usage{
\S4method{oracle}{Cat}(catObj, theta, ans_profiles)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}

\item{theta}{A numeric representing the true position on the latent trait.}

\item{ans_profiles}{A vector representing the respondent's full answer profile.}
}
\value{
A data.frame where the first column is the user-supplied true value of theta, the second column is the
best possible theta estimate given n questions are asked, and the remaining columns are the answer profile leading
to the best possible theta estimation.
}
\description{
Generating all possible combinations of length n from a response profile to determine
the possible response profile best able to estimate the true value of theta.
}
\details{
lengthThreshold slot should specify how many questions to ask.
}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery, Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
