% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Cat.R
\docType{methods}
\name{plot.Cat}
\alias{plot.Cat}
\alias{plot,Cat}
\alias{plot,Cat-method}
\title{Plotting function for Cat object}
\usage{
\S4method{plot}{Cat}(x, item = "numeric", type = "character", ...)
}
\arguments{
\item{x}{Cat.}

\item{item}{Numeric.}

\item{type}{Character.}

\item{...}{Other arguments passed to plot().}
}
\description{
Description.
}
