% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polknowMT}
\alias{polknowMT}
\title{MTurk Political Knowledge}
\format{
A data frame with 810 observations on the following 64 questions.
\describe{

\item{\code{Q1}}{How long is one term for the President of the United States?

Eight years; Six years; Four years; Two years}

\item{\code{Q2}}{The FDA is part of the national government primarily responsible for regulating
  
  Food quality; The national parks; Electricity production and energy; Pollution and the environment}

\item{\code{Q3}}{Who is the Vice President of the United States?
  
  Leon Panetta; William Daley; Hillary Clinton; Joe Biden}

\item{\code{Q4}}{The federal debt is 
  
  Much smaller than it was 20 years ago; The difference between imports and exports with foreign countries; 
  The annual difference between spending and tax revenues; The accumulated borrowing 
  of the federal government that has not been repaid}

\item{\code{Q5}}{How many times can an individual be elected President of the United States under current laws?
  
  Any number of terms; Three times; Twice; Once}

\item{\code{Q6}}{What do we call the first ten amendments to the Constitution?
  
  The Articles of Confederation; The inalienable right; The Bill of Rights; The Declaration of Independence}

\item{\code{Q7}}{Is the U.S. federal budget deficit, the amount by which the governments 
spending exceeds the amount of money it collects, now bigger, about the same, or smaller 
than it was during most of the 1990s?
  
  Smaller; About the same; Bigger}

\item{\code{Q8}}{Who signs bills to become laws?
  
  The President; The Vice President; The Chief Justice of the Supreme Court; The Secretary of State}

\item{\code{Q9}}{Which party is generally more supportive of creating a way for immigrants who 
are in the U.S. illegally to eventually become citizens?
  
  The Republican Party; The Democratic Party}

\item{\code{Q10}}{In what month do we vote for the President?
  
  November; October; February; January}

\item{\code{Q11}}{What are the two parts of the U.S. Congress?
  
  The Senate and the Supreme Court; The House of Lords and the House of Commons; The House of 
  Representative and the Supreme Court; The Senate and House of Representatives}

\item{\code{Q12}}{Which party is generally more supportive of restricting access to abortion?
  
  The Republican Party; The Democratic Party}

\item{\code{Q13}}{Which of these political parties is considered most conservative?
  
  Green Party; Republican Party; Democratic Party}

\item{\code{Q14}}{The NRA is an organization that advocates for 
  
  Clean elections; A cleaner environment; The rights of gun owners; Women's rights}
  
  \item{\code{Q15}}{Compared to 30 years ago, is the difference in incomes between the top 
  20\% of households and the bottom 20\% of households now bigger, smaller, or the same?
  
  The difference is now the same as 30 years ago; The difference is now smaller than 30 years ago; 
  The difference is now bigger than 30 years ago}
  
  \item{\code{Q16}}{The EPA is part of the national government primarily responsible for regulating
  
  Food quality; The national parks; Electricity production and energy; Pollution and the environment}
  
  \item{\code{Q17}}{Which party is generally more supportive of reducing the size of the defense budget?
  
  The Republican Party; The Democratic Party}
  
  \item{\code{Q18}}{Which party is generally more supportive of increasing taxes on higher 
  income people to reduce the federal budget deficit?
  
  The Republican Party; The Democratic Party}
  
  \item{\code{Q19}}{Which party is generally more supportive of allowing drilling for oil in 
  the Arctic National Wildlife Refuge?
  
  The Republican Party; The Democratic Party}
  
  \item{\code{Q20}}{Who is the Commander in Chief of the military?
  
  The Attorney General; The President; The Secretary of Defense; The Vice President}
  
  \item{\code{Q21}}{The First Amendment to the United States Constitution guarantees all of these rights EXCEPT
  
  Right to remain silent; Right to the free exercise of religion; Right to free speech; Right to peaceably assemble}
  
  \item{\code{Q22}}{Roe v Wade is a case decided by the Supreme Court that relates to
  
  Executive power; Campaign finance; Birth control; Abortion}
  
  \item{\code{Q23}}{Social Security is
  
  Funded by the personal income tax; Operated by state government; The responsibility of the 
  Department of Defense; The benefit program for senior citizens}
  
  \item{\code{Q24}}{What is Medicare?
  
  A private, non-profit organization that runs free health clinic; A private health insurance 
  plan sold to individuals in all 50 states; A program run by state governments to provide health 
  care to poor people; A program run by the U.S. federal government to pay for old people's health care}

\item{\code{Q25}}{How many senators are elected from each state?
  
  It depends on the population of the state; Four; Two; One}

\item{\code{Q26}}{How many votes are required in Congress to override a presidential veto
  
  A simple majority of both houses of Congress; A simple majority of one house of Congress; 
  A two-thirds majority of both houses of Congress; A two-thirds majority of one house of Congress}

\item{\code{Q27}}{The Secretary of State
  
  Serves a two-year term; Serves the state governments; Is nominated by the president; Heads 
  the armed services}

\item{\code{Q28}}{Near the end of an election campaign, a polls shows that an issue that no 
candidate has mentioned is of great concern to voters. What is most likely to happen?
  
  Some candidates will drop out of the race; Candidates will start talking about the issue; 
  Newspapers will not report the results of the poll; The election will be postponed}

\item{\code{Q29}}{Liberals are generally said to
  
  Support pro-life policies; Oppose all tax increases; Support military spending; Support 
  government programs to give government assistance the needy}

\item{\code{Q30}}{Which party is generally more supportive of reducing the size and scope 
of the federal government?
  
  The Republican Party; The Democratic Party}

\item{\code{Q31}}{The ability of a minority of senators to prevent a vote on a bill 
is known as Suspension of the rules; Enrollment; A veto; A filibuster}

\item{\code{Q32}}{conservatives are generally said to
  
  Support pro-choice policies; Support tax cuts; Oppose military spending; 
  Support government programs to give government assistance to the needy}

\item{\code{Q33}}{Which of these regions of the country is generally 
considered to be most supportive of Republican candidates 
  
  Midwest; South; West Coast; New England}

\item{\code{Q34}}{The presiding officer in the House of Representatives is 
  
  The Majority Leader; The Sergeant at Arms; The Vice President of the 
  United States; The Speaker}

\item{\code{Q35}}{Which of these countries is NOT a permanent member of 
the U.N. Security Council
  
  United Kingdom; France; India; China}

\item{\code{Q36}}{Which part has a majority of seats in the U.S. House 
of Representatives?
  
  Neither; Democrats; Republicans}

\item{\code{Q37}}{What state holds the first primary election in 
Presidential primaries?
  
  Florida; Nevada; South Carolina; New Hampshire}

\item{\code{Q38}}{Who is the Speaker of the House of Representatives?
  
  Mitt Romney; Eric Holder; Harry Reid; John Boehner}

\item{\code{Q39}}{Most cases are considered by the Supreme Court
  
  In even-numbered years; At the request of the Congress; Upon order 
  of the president; With the approval of at least four justices}

\item{\code{Q40}}{How many Justices typically serve on the U.S. Supreme Court
  
  Eleven; Nine; Eight; Seven}

\item{\code{Q41}}{What job or political office is no held by Ben Bernanke?
  
  None of these; Minority Whip of the U.S. House; Chief Justice of 
  the United States Supreme Court; Majority leader of the U.S. Senate, Chairman of the Federal Reserve}

\item{\code{Q42}}{Whose responsibility is it to nominate judges to the Federal Courts
  
  The state governors; The Supreme Court; Congress; President}

\item{\code{Q43}}{Who is the Chief Justice of the U.S. Supreme Court?
  
  Larry Thompson; Anthony Kennedy; David Cole; John Roberts}

\item{\code{Q44}}{The U.S. Senate 
  
  Votes to confirm nominees to the U.S. Supreme court chosen by the 
  House of Representatives; Plays no role in choosing the members 
  of the U.S. Supreme Court; Chooses members of the U.S. Supreme 
  Court; Votes to confirm nominees to the U.S. Supreme Court chosen by the President}

\item{\code{Q45}}{Which party has a majority of seats in the U.S. Senate
  
  Neither; Democrats; Republicans}

\item{\code{Q46}}{Which of the states listed below has the 
greatest number of electoral college votes in the U.S. Presidential Elections?
  
  Puerto Rico; Nevada; North Dakota; Washington, D.C.}

\item{\code{Q47}}{Citizens United v the FEC is a case decided 
by the Supreme Court that relates to
  
  Executive power; Campaign finance; Birth control; Abortion}

\item{\code{Q48}}{For how many years is a United States Senator 
elected that is, how many years are there in one full term of office for a U.S. Senator?
  
  None of these; Eight years; Six years; Four years; Two years}

\item{\code{Q49}}{Who is the Prime Minister of the United Kingdom?
  
  Richard Branson; Tony Hayward; Nick Clegg; David Cameron}

\item{\code{Q50}}{The president of Afghanistan is named
  
  Bashar al-Assad; Hosni Mubarak; Hamid Karzai; Nouri al-Maliki}

\item{\code{Q51}}{The House of Representatives has how many voting members?
  
  Four hundred and forty-one; Four hundred and thirty-five; 
  Two hundred; One hundred}

\item{\code{Q52}}{The President of the Senate is
  
  The Majority Leader; The Sergeant at Arms; The Vice President 
  of the United States; The senior senator of the majority party}

\item{\code{Q53}}{On which of the following federal programs is 
the most money spent each year?
  
  Medicare; Education; Subsidies to farmers; Aid to foreign countries}

\item{\code{Q54}}{What do all constitutional governments have?
  
  Separation of church and state; A bill of rights; A President 
  as the head of government; Limits on political power}

\item{\code{Q55}}{One which of the following does the U.S. 
federal government spend the least money?
  
  Social Security; National defense; Medicare; Foreign aid}

\item{\code{Q56}}{The head of the Department of Justice is
  
  Kathleen Sebelius; Eric Holder; Timothy Geithner; Hillary Clinton}

\item{\code{Q57}}{The president may NOT
  
  Declare war; Pardon criminals without justification; Appoint 
  federal officials when Congress is in recess; Refuse to sign 
  legislation passed by Congress}

\item{\code{Q58}}{Which of these is NOT primarily the 
responsibility of the Federal government in Washington?
  
  Interstate commerce; Negotiating treaties with foreign countries; 
  Education; National defense}

\item{\code{Q59}}{Who is the current president of Mexico?
  
  Vincente Fox; Hugo Chavez; Dilma Rousseff; Felipe Calderon}

\item{\code{Q60}}{Which of the following actions does the United 
States federal government commonly take to finance a budget deficit?
  
  Expanding public-works projects; Borrowing from the public; 
  Imposing import quotas; Printing more money}

\item{\code{Q61}}{Common Cause is an organization that advocates for
  
  Women's rights; Clean elections; A cleaner environment; The right of gun owners}
  
  \item{\code{Q62}}{The Byrd Rule is relevant
  
  During the confirmation of cabinet members; For national party conventions; 
  During Congressional debates over non-budgetary policies; For the 
  Reconciliation process}
  
  \item{\code{Q63}}{The Majority Leader of the House of Representative is
  
  Nancy Pelosi; Kevin McCarthy; Eric Cantor; John Boehner}
  
  \item{\code{Q64}}{On which of the following does the U.S. federal 
  government spend the most money each year?
  
  Education; Medicare; Interest on the national debt; National defense}
}
}
\source{
\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=hdl:1902.1/19381}
}
\usage{
data(polknowMT)
}
\description{
Data from Amazon Mechanical Turkers in 2012 responding to 64 political knowledge questions.  
Political knowledge questions had a varying number of response options, noted below.  
This dataset codes answers as either correct (a value of 0) or incorrect (a value of 1).
}
\references{
Jacob M. Montgomery and Joshua Cutler. 2013. 
"Computerized Adaptive Testing for Public Opinion Surveys." 
Political Analysis (Spring 2013) 21 (2): 172-192.
}
\keyword{datasets}
