% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polknowOrdered}
\alias{polknowOrdered}
\title{TAPS Political Knowledge (Ordered Response Options)}
\format{A data frame with 1340 observations (only complete cases) of the following 10 questions.
\describe{

\item{\code{Q1}}{Members of the U.S. Supreme Court serve 

3 = two-year terms. 4 = ten-year terms. 5 = life terms. 2 = terms determined by the president. 1 = Don't know.
}

\item{\code{Q2}}{Who is the Chief Justice of the U.S. Supreme Court?

5 = John Roberts. 4 = Antonin Scalia. 3 = Mitt Romney. 2 = Hillary Clinton. 1 = Don't know.
}

\item{\code{Q3}}{Social Security is

5 = the benefit program for senior citizens. 2 = the responsibility of the 
Department of Defense. 3 = operated by state governments. 4 = funded by the personal income tax. 1 = Don't know.
}

\item{\code{Q4}}{On which of the following programs is the most money spent each year?

4 = aid to foreign countries. 5 = Medicare. 2 = subsidies to farmers. 3 = education. 1 = Don't know.
}

\item{\code{Q5}}{Which party holds a majority of seats in the U.S. House of Representatives in Washington?

4 = Democrats. 3 = Republicans. 2 = Independents. 1 = Don't know.
}

\item{\code{Q6}}{How many votes are required in Congress to override a presidential veto?

2 = a simple majority of one house of Congress. 3 = a simple majority of both 
houses of Congress. 4 = a two-thirds majority of one house of Congress. 5 = a 
two-thirds majority of both houses of Congress. 1 = Don't know.
}

\item{\code{Q7}}{How long is one term for a member of the U.S. Senate?

3 = two years. 4 = four years. 5 = six years. 2 = eight years. 1 = Don't know.
}

\item{\code{Q8}}{The ability of a minority of senators to prevent a vote on a bill is known as 

4 = a veto. 5 = a filibuster. 3 = enrollment. 2 = suspension of the rules. 1 = Don't know.
}

\item{\code{Q9}}{Who is the Vice President of the United States?

4 = Nancy Pelosi. 3 = John Boehner. 5 = Joseph Biden. 2 = Harry Reid. 1 = Don't know.
}

\item{\code{Q10}}{A president may serve

3 = one term. 5 = two terms. 4 = three terms. 2 = any number of terms. 1 = Don't know.
}
}}
\source{
\url{https://wc.wustl.edu/}
}
\usage{
data(polknowOrdered)
}
\description{
Data of responses to a political knowledge battery from the May 2013 wave of 
The American Panel Survey (TAPS) out of Washington University in St. Louis.
}
\details{
TAPS is a monthly online panel survey  of about 2,000 adults in the United 
States. The panel was recruited in the fall of 2011 using an address-based 
sampling frame.  TAPS surveys are administered online.  Selected panelists 
who do not have a computer or online service are provided a computer and 
internet access by TAPS.
}
\keyword{datasets}
