% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prior}
\alias{prior}
\title{The prior value for the respondent's position on the latent trait scale}
\usage{
prior(x, c, p)
}
\arguments{
\item{x}{A numeric value at which to evaluate the prior}

\item{c}{The \code{priorName} slot of a \code{Cat} object; indicates the distribution}

\item{p}{The \code{priorParams} slot of a \code{Cat} object}
}
\value{
A vector consisting of prior value, \eqn{\pi(x)}, given the value \eqn{x}
}
\description{
Calculates the prior value for a respondent's position on the latent scale of interest
}
\details{
Note: \eqn{x} needs to be either UNIFORM, NORMAL, or STUDENT_T parameters, which control the shape of the prior.
}
\examples{

 
 
}
\seealso{
\code{\link{Cat}} for additional information on priors; 
 \code{\link{dLL}} and/or \code{\link{d2LL}} for application of priors
}

