% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processAJAX.R
\docType{methods}
\name{processAJAX}
\alias{processAJAX}
\alias{processAJAX,character-method}
\title{Qualtrics AJAX Handler}
\usage{
\S4method{processAJAX}{character}(catObj, item)
}
\arguments{
\item{catObj}{An object of class \code{character}.}

\item{item}{An integer indicating the index of the question item}

\item{answer}{The answer to the \code{item} to be updated}
}
\description{
Qualtrics AJAX Handler used to implement catSurv functionality in a Qualtrics survey
}
\author{
Joshua Landman
}
