% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{selectItem}
\alias{selectItem}
\title{Select Next Item}
\usage{
selectItem(catObj)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}
}
\value{
The function \code{selectItem} returns a \code{list} with two elements:
 
\code{estimates}: a \code{data.frame} with a row for each unasked question and three columns representing 
the item index number, the item name, and the item value (calculated by the specified selection method), 
and

\code{next_item}: a numeric representing the index of the item that should be asked next.
}
\description{
Selects the next item in the question set to be adminstered to respondent based on the specified selection method.
}
\details{
Selection approach is specified in the \code{selection} slot of the \code{Cat} object.

The minimum expected posterior variance criterion is used when the \code{selection}
 slot is \code{"EPV"}.

The maximum Fisher's information criterion is used when the \code{selection}
  slot is \code{"MFI"}.
  
The maximum likelihood weighted information criterion is used when the \code{selection}
slot is \code{"MLWI"}.

The maximum posterior weighted information criterion is used when the \code{selection}
slot is \code{"MPWI"}.
 
The maximum expected information criterion is used when the \code{selection}
slot is \code{"MEI"}.

The maximum Kullback-Leibler information criterion is used when the \code{selection}
slot is \code{"KL"}.

The maximum likelihood weighted Kullback-Leibler information criterion is used when the \code{selection}
slot is \code{"LKL"}.

The maximum posterior weighted Kullback-Leibler information criterion is used when the \code{selection}
slot is \code{"PKL"}.

The ??????????? criterion is used when the \code{selection}
slot is \code{"MFII"}.

A random number generator is used when the \code{selection}
slot is \code{"RANDOM"}.
}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled C++ code.

The \code{"RANDOM"} item selection criterion uses the package \code{RcppArmadillo} to randomly
choose the next item among unasked questions.  \code{RcppArmadillo} provides an exact reproduction
of R's \code{sample} function that can be called from C++.
}
\examples{
\dontrun{
## Create Cat object
data(npi)
ltm_cat <- ltmCat(npi)

## Store example answers
setAnswers(ltm_cat) <- c(1,0,1,0,1, rep(NA, 35))

## Set different selection criterion and choose next item
setSelection(ltm_cat) <- "EPV"
selectItem(ltm_cat)

setSelection(ltm_cat) <- "MFI"
selectItem(ltm_cat)

setSelection(ltm_cat) <- "MLWI"
selectItem(ltm_cat)

setSelection(ltm_cat) <- "MPWI"
selectItem(ltm_cat)

setSelection(ltm_cat) <- "MEI"
selectItem(ltm_cat)

setSelection(ltm_cat) <- "KL"
selectItem(ltm_cat)

setSelection(ltm_cat) <- "LKL"
selectItem(ltm_cat)

setSelection(ltm_cat) <- "PKL"
selectItem(ltm_cat)

setSelection(ltm_cat) <- "MFII"
selectItem(ltm_cat)

setSelection(ltm_cat) <- "RANDOM"
selectItem(ltm_cat)
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{estimateTheta}}, \code{\link{expectedPV}}, \code{\link{fisherInf}}
}

