% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectItem_qualtrics.R
\name{selectItem_qualtrics}
\alias{selectItem_qualtrics}
\title{Convert JSON/Character object to Cat object and calls SelectItem}
\usage{
selectItem_qualtrics(jsonObj)
}
\arguments{
\item{jsonObj}{A \code{json} or character object with elements matching Cat object slots}
}
\value{
Results from \code{selectItem}
}
\description{
Convert JSON/Character object to Cat object and calls SelectItem
}
\seealso{
\code{\link{Cat-class}}
}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery, Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
