% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cat-class.R
\name{setGuessing<-}
\alias{setGuessing<-}
\title{Setter for guessing slot within Cat class objects}
\usage{
setGuessing(object, valid = T) <- value
}
\arguments{
\item{object}{An object of class \code{Cat} to be updated}

\item{valid}{Boolean for whether to check validity of object before allowing setter method to execute, default to TRUE}

\item{value}{The value to replace the current value stored in the guessing slot of the \code{Cat}}
}
\value{
Updates the \code{Cat} object that was provided by the user, and returns the updated object
}
\description{
This functions replaces the values currently stored in the guessing slot of an object of class \code{Cat} with the user-provided values.
}
\note{
There are setters for every slot in \code{Cat} objects, all following the format \code{setSlot}
}
\author{
Josh W. Cutler: \email{josh@zistle.com} and Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
\seealso{
\code{\link{Cat}}, \code{\link{setDiscrimination}}, \code{\link{setPriorParam}}
}

