% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cat-class.R
\name{setguessing<-}
\alias{setguessing<-}
\title{Setters for slots within Cat class objects}
\usage{
setguessing(object) <- value
}
\arguments{
\item{object}{An object of class \code{Cat} with a slot to be changed}

\item{value}{The value to replace the current value stored in the slot of the \code{Cat}, as specified in the function name

 @return An object of class \code{Cat} with an updated slot, specified in the function name}
}
\description{
These functions replace the values currently stored in a given slot of an object of class \code{Cat} with the user-provided values.
}
\note{
Notes if you need
}
\author{
Josh W. Cutler: \email{josh@zistle.com} and Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
\seealso{
\code{\link{Cat}}
}

