% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/three.pl.R
\name{three.pl}
\alias{three.pl}
\title{Computerized Adaptive Testing Survey Probability Function}
\usage{
three.pl(cat, theta, difficulty, discrimination, guessing)
}
\arguments{
\item{cat}{an object of \code{CATsurv} class.}

\item{theta}{vector consisting of each respondent's position on the latent scale of interest.}

\item{difficulty}{vector consisting of difficulty parameter for each item.}

\item{discrimination}{vector consisting of discrimination parameter for each item.}

\item{guessing}{vector consisting of guessing parameter for each item .}
}
\value{
A vector consisting of the probability of a correct response for each respondent on item \eqn{i}.
}
\description{
This function returns the probability of a correct response for each respondent on item \eqn{i}.
}
\details{
The probability of a correct response for respondent \eqn{j} on item \eqn{i} is \deqn{Pr(y_{ij}=1|\theta_j)=c+(1-c)\frac{\exp(Da_i(\theta_j-b_i))}{1+\exp(Da_i(\theta_j-b_i))}} where \eqn{\theta_j} is respondent \eqn{j}'s position on the latent scale of interest, \eqn{a_i} is item \eqn{i}'s discrimination parameter, \eqn{b_i} is item \eqn{i}'s difficulty parameter, and \eqn{c_i} is item \eqn{i}'s guessing parameter.
}
\author{
Josh W. Cutler and Jacob M. Montgomery
}
\seealso{
\code{\link{likelihood}},\code{\link{prior.value}}, \code{\link{estimateTheta}}, \code{\link{estimateSE}}, \code{\link{expectedPV}}, \code{\link{nextItem}}, \code{\link{storeAnswer}}, \code{\link{debugNextItem}}
}

