% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toJSONCat.R
\name{toJSONCat}
\alias{toJSONCat}
\alias{toJSONCat,Cat-method}
\title{Convert Cat object to JSON}
\usage{
\S4method{toJSONCat}{Cat}(catObj)
}
\arguments{
\item{catObj}{A \code{Cat} object}
}
\value{
The function \code{toJSONCat} returns a JSON object with list elements corresponding to information stored in the \code{catObj}.

See \code{\link{Cat-class}} for required \code{Cat} object slots.
}
\description{
This function  object of class \code{Cat}.
}
\seealso{
\code{\link{Cat-class}}
}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery, Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
