% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tpm_cat}
\alias{tpm_cat}
\title{tpm Cat Object}
\format{
An object of class \code{Cat}.  See \code{\link{Cat-class}} for more details.
}
\usage{
data(tpm_cat)
}
\description{
An object of class \code{Cat} created using the \code{tpmCat} function with the first twenty questions of the \code{polknowMT} dataset.
}
\examples{
\dontrun{
## How this Cat object was created
data(polknowMT)
tpm_cat <- tpmCat(polknowMT[,1:20], quadraturePoints = 100, start.val = "random")
}

## How to load this Cat object for usage
data(tpm_cat)

}
\seealso{
\code{\link{Cat-class}}, \code{\link{tpmCat}}, \code{\link{polknowMT}}
}
\keyword{datasets}
