% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{checkStopRules}
\alias{checkStopRules}
\title{Check if Stop and/or Override Rules are Met}
\usage{
checkStopRules(catObj)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}
}
\value{
This function returns a boolean, where TRUE indicates the the stopping rules are met (no further questions should be asked) and FALSE indicates the stoppings rules are not met (additional questions are needed)
}
\description{
Evaluates the specified stopping and/or override rules to check if respondent should be asked further questions
}
\details{
The stopping rule thresholds are stored in the following Cat object slots: lengthThreshold, seThreshold, infoThreshold, and gainThreshold.
  The override thresholds are stored in the following Cat object slots: lengthOverride, gainOverride. A value of NA indicates the rule should not be used.

  A return value of TRUE indicates that additional questions should be asked; FALSE indicates no additional questions should be asked.

  A user can specify any combination of stopping rules and/or overrides.  The function returns TRUE if all specified stopping rules are met
  and no specified overrides are met. The function returns FALSE if at least one specified stopping rule is not met, or if any specified override threshold is met.

  Stopping Rules:

  lengthThreshold: Number of question's answered >= a
 
  seThreshold: \eqn{SE(\hat{\theta}) < a}
 
  infoThreshold: \deqn{FI < a \forall}{FI < a \forall} remaining items

  gainThreshold: \deqn{SE(\hat{\theta}) - \sqrt{EPV} | < a \forall}{SE(\hat{\theta}) - \sqrt{EPV} | < a \forall} remaining items

  Overrides:

  lengthOverride: Number of question's answered < a

  gainOverride: \deqn{|SE(\hat{\theta}) - \sqrt{EPV} | >= a \forall}{|SE(\hat{\theta}) - \sqrt{EPV} | >= a \forall} remaining items
}

