% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimateSE}
\alias{estimateSE}
\title{Standard Error of Ability Parameter Estimate}
\usage{
estimateSE(catObj)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}
}
\value{
The function \code{estimateSE} returns a numeric for the standard error for \eqn{\theta}.
}
\description{
Estimates the standard error for a respondent's ability parameter estimate, \eqn{\theta}.
}
\details{
Estimation approach is specified in \code{estimation} slot of \code{Cat} object.  The
options are \code{"EAP"} for the expected a posteriori approach, \code{"MAP"} for the modal a posteriori
approach, \code{"MLE"} for the maximum likelihood approach, and \code{"WLE"} for the weighted maximum likelihood
approach.  The function \code{estimateSE} will calculate the standard error of the ability estimate
given the estimation approach.
}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled C++ code.
}
\examples{
\dontrun{
## Create Cat object
data(npi)
ltm_cat <- ltmCat(npi)

## Store example answers
setAnswers(ltm_cat) <- c(1,0,1,0,1, rep(NA, 35))

## Set different estimation procedures and calculate
## ability estimate and its standard error
setEstimation(ltm_cat) <- "EAP"
estimateTheta(ltm_cat)
estimateSE(ltm_cat)

setEstimation(ltm_cat) <- "MAP"
estimateTheta(ltm_cat)
estimateSE(ltm_cat)

setEstimation(ltm_cat) <- "MLE"
estimateTheta(ltm_cat)
estimateSE(ltm_cat)

setEstimation(ltm_cat) <- "WLE"
estimateTheta(ltm_cat)
estimateSE(ltm_cat)
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{estimateTheta}}
}

