% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimateSE}
\alias{estimateSE}
\title{Estimate of standard error for the posterior estimate}
\usage{
estimateSE(cat_df)
}
\arguments{
\item{cat}{An object of \code{Cat} class}
}
\value{
The posterior standard error for \eqn{\theta_j}
}
\description{
This function estimates the standard error for the posterior estimate for the position \eqn{\theta_j} of a person on the latent scale
}

