% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimateTheta}
\alias{estimateTheta}
\title{Estimate of the respondent's ability parameter}
\usage{
estimateTheta(catObj)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}
}
\value{
A numeric consisting of the expected value of the ability parameter
}
\description{
Calculates the expected value of the ability parameter \eqn{\theta}, conditioned on the observed answers and the item calibrations
}
\details{
EAP
MAP (This method is only available using the normal prior distribution.)
The maximum likelihood (MLE) approach: (Note: When MLE will not work, \code{estimateTheta} is calculated according to the \code{estimationDefault} slot in \code{Cat} object.)
The weighted maximum likelihood approach (WLE) approach: Estimating \eqn{\theta_j} requires finding the root of the following function using the ``Brent'' method in the \code{gsl} library.
}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled C++ code.
}
\examples{
\dontrun{
## Prior calculation using Cat object of the ltm model
## specifying different distributions
data(npi)
cat <- ltmCat(npi)
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{probability}} and \code{\link{likelihood}}
}

