% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimateTheta.R
\name{estimateTheta}
\alias{estimateTheta}
\title{Computerized Adaptive Testing Survey Estimating Each Respondent's Position on the Latent Scale of Interest}
\usage{
estimateTheta(cat, ...)
}
\arguments{
\item{cat}{an object of class \code{CATsurv}.}

\item{...}{argument passed to other functions.}
}
\value{
A vector consisting of the expected \emph{a posteriori} estimate for each respondent's position on the latent scale of interest.
}
\description{
This function returns the expected \emph{a posteriori} (EAP) estimate of each respondent's popsition on the latent scale.
}
\details{
The expected \emph{a posteriori} (EAP) estimate of respondent \eqn{j}'s position on the latent scale is calculated as \deqn{\hat{\theta}_j^{(\text{EAP})}\equiv E(\theta_j|\mathbf{y}_{k-1j})=\frac{\int\theta_j\pi(\theta_j)L(\theta_j|\mathbf{y}_{k-1j}d\theta_j}{\int\pi(\theta_j)L(\theta_j|\mathbf{y}_{k-1j}d\theta_j}}.
}
\author{
Josh W. Cutler and Jacob M. Montgomery
}
\seealso{
\code{\link{three.pl}},\code{\link{likelihood}}, \code{\link{prior.value}}, \code{\link{estimateSE}}, \code{\link{expectedPV}}, \code{\link{nextItem}}, \code{\link{storeAnswer}}, \code{\link{debugNextItem}}
}

