% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expectedObsInf}
\alias{expectedObsInf}
\title{Expected Observed Information}
\usage{
expectedObsInf(cat_df, item)
}
\arguments{
\item{cat_df}{An object of class \code{Cat}}

\item{item}{An integer indicating the index of the question}
}
\value{
A value of the expected information
}
\description{
This function calculates the expected information, which is the observed information attained from a specific response set times the probability of that profile occurring.
}
\details{
Binary details:

  \deqn{EI=P(y_{ij} = 1)J_{y_{ij}=1} + P(y_{ij} = 0)J_{y_{ij}=0}}{EI = P(y_ij = 1)J_{y_ij = 1} + P(y_ij = 0) J_{y_ij = 0}} 

  Categorical details:

  \deqn{EI = \sum_k^K P(y_{ij} = k)J_{y_{ij}=k}}{EI = \sum_k^K P(y_ij = k) J_{y_ij = k}}
}
\examples{

 
 
}
\seealso{
\code{\link{probability}} and/or \code{\link{likelihood}} for calculation of P and P*;  
  \code{\link{estimateTheta}} for calculation of \eqn{\theta};  
  \code{\link{obsInf}} for observed information calculation
}

