% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expectedPV}
\alias{expectedPV}
\title{Expected Posterior Variance}
\usage{
expectedPV(cat_df, item)
}
\arguments{
\item{cat_df}{An object of class \code{Cat}}

\item{item}{An integer indicating the index of the question}
}
\value{
A numeric value indicating a respondent's expected posterior variance
}
\description{
This function estimates the expected posterior variance for a respondent's estimated position on the latent trait for an item yet to be answered based on a respondent's position on the latent trait from the already-answered items.
}
\details{
For a binary question:

  \deqn{P(y_{ij} = 1 | \hat{\theta})Var(\hat{\theta} | y_{ij} = 1) + P(y_{ij} = 0 | \hat{\theta})Var(\hat{\theta} | y_{ij} = 0)}{P(y_ij = 1 | \hat{\theta}) Var(\hat{\theta} | y_ij = 1) + P(y_ij = 0 | \hat{\theta}) Var(\hat{\theta} | y_ij = 0)}

  For a categorical question:

  \deqn{\sum_{i=k}^k P(y_{ij} = k | \hat{\theta})Var(\hat{\theta} | y_{ij} = k)}{\sum_{i = k}^k P(y_ij = k | \hat{\theta}) Var(\hat{\theta} | y_ij = k)}
}
\examples{


}
\seealso{
\code{\link{probability}} for calculation of P;  
  \code{\link{estimateTheta}} for calculation of \eqn{\theta}
}

