% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expectedPV}
\alias{expectedPV}
\title{Expected Posterior Variance}
\usage{
expectedPV(cat_df, item)
}
\arguments{
\item{cat_df}{An object of \code{Cat} class}

\item{item}{An integer indicating the index of the question}
}
\value{
A numeric value indicating a respondent's expected posterior variance
}
\description{
This function estimates the expected posterior variance for a respondent's estimated position on the latent trait for an item yet to be answered based on a respondent's position on the latent trait from the already-answered items.
}
\details{
For a binary question:

  \deqn{P(y_{ij} = 1 | \hat{\theta})Var(\hat{\theta} | y_{ij} = 1) + P(y_{ij} = 0 | \hat{\theta})Var(\hat{\theta} | y_{ij} = 0)}{P(y_ij = 1 | \hat{\theta}) Var(\hat{\theta} | y_ij = 1) + P(y_ij = 0 | \hat{\theta}) Var(\hat{\theta} | y_ij = 0)}

  For a categorical question:

  \deqn{\sum_{i=k}^k P(y_{ij} = k | \hat{\theta})Var(\hat{\theta} | y_{ij} = k)}{\sum_{i = k}^k P(y_ij = k | \hat{\theta}) Var(\hat{\theta} | y_ij = k)}
}
\examples{

## binary (ltm)

 data("npi")
 ltm_data <- npi[1:100, ]
 ltm_cat <- ltmCat(ltm_data, quadraturePoints = 100)
 ltm_cat@answers <- as.numeric(ltm_data[1,])
 
 expectedPV(ltm_cat, 1)
 
## binary (tpm)

 data("AMTknowledge")
 tpm_data <- AMTknowledge[1:100, ]
 tpm_cat <- tpmCat(tpm_data, quadraturePoints = 100)
 tpm_cat@answers <- as.numeric(tpm_data[1,])
 
 expectedPV(tpm_cat, 1)

## categorical (grm) 

 data("nfc")
 poly_data <- nfc[1:100, ]
 poly_cat <- grmCat(poly_data, quadraturePoints = 100)
 poly_cat@answers <- as.numeric(poly_data[1,])
 
 expectedPV(poly_cat, 1)

}
\seealso{
\code{\link{probability}} for calculation of P;  
  \code{\link{estimateTheta}} for calculation of \eqn{\theta}
}

