% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fisherInf}
\alias{fisherInf}
\title{Fisher's Information}
\usage{
fisherInf(catObj, theta, item)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}

\item{theta}{A numeric or an integer indicating the potential value for \eqn{\theta_j}}

\item{item}{An integer indicating the index of the question item}
}
\value{
The function returns a numeric of the expected value of the observed information
}
\description{
Calculates the expected value of the observed information of the likelihood evaluated at the input value \eqn{\theta}
}
\details{
For the dichotomous case, this is equivalent to the observed information.
}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled C++ code.
}
\examples{
\dontrun{
## Prior calculation using Cat object of the ltm model
## specifying different distributions
data(npi)
cat <- ltmCat(npi)
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{obsInf}} for observed information calculation and
  \code{\link{fisherTestInfo}} for further application of Fisher's information
}

