% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fisherTestInfo}
\alias{fisherTestInfo}
\title{Fisher's Test Information}
\usage{
fisherTestInfo(catObj)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}
}
\value{
The total information gained for a respondent, given a specific answer set and a value of \eqn{theta}.
}
\description{
This function calculates the total information gained for a respondent \eqn{j} for all answered items, conditioned on \eqn{theta}.
}
\details{
\deqn{I_i(\theta_j)}{I_i(\theta_j)} is the Fisher's information for item \eqn{i} specific to \eqn{\theta_j} (see \code{FisherInf}). 
  Let \deqn{i\in[1, \ldots, n]}{i \in [1,...,n]} index the questions the respondent has already answered. The test information is then defined as:

\deqn{I(\theta_j) = \sum_i^n I_i(\theta_j)}{I(\theta_j) = \sum_i^n I_i(\theta_j)}
}
\examples{

 

}
\seealso{
\code{\link{probability}} and/or \code{\link{likelihood}} for calculation of P and P*;  
  \code{\link{estimateTheta}} for calculation of \eqn{\theta};
  \code{\link{fisherInf}} for further calculation of Fisher's information
}

