% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cat-class.R
\docType{methods}
\name{getters}
\alias{getAnswers}
\alias{getAnswers,Cat-method}
\alias{getDifficulty}
\alias{getDifficulty,Cat-method}
\alias{getDiscrimination}
\alias{getDiscrimination,Cat-method}
\alias{getEstimation}
\alias{getEstimation,Cat-method}
\alias{getEstimationDefault}
\alias{getEstimationDefault,Cat-method}
\alias{getGainOverride}
\alias{getGainOverride,Cat-method}
\alias{getGainThreshold}
\alias{getGainThreshold,Cat-method}
\alias{getGuessing}
\alias{getGuessing,Cat-method}
\alias{getInfoThreshold}
\alias{getInfoThreshold,Cat-method}
\alias{getLengthOverride}
\alias{getLengthOverride,Cat-method}
\alias{getLengthThreshold}
\alias{getLengthThreshold,Cat-method}
\alias{getLowerBound}
\alias{getLowerBound,Cat-method}
\alias{getModel}
\alias{getModel,Cat-method}
\alias{getPriorName}
\alias{getPriorName,Cat-method}
\alias{getPriorParams}
\alias{getPriorParams,Cat-method}
\alias{getSeThreshold}
\alias{getSeThreshold,Cat-method}
\alias{getSelection}
\alias{getSelection,Cat-method}
\alias{getUpperBound}
\alias{getUpperBound,Cat-method}
\alias{getZ}
\alias{getZ,Cat-method}
\alias{getters}
\title{Methods for Getting Slot from Cat Object}
\usage{
\S4method{getModel}{Cat}(object)

\S4method{getGuessing}{Cat}(object)

\S4method{getDiscrimination}{Cat}(object)

\S4method{getDifficulty}{Cat}(object)

\S4method{getAnswers}{Cat}(object)

\S4method{getPriorName}{Cat}(object)

\S4method{getPriorParams}{Cat}(object)

\S4method{getLowerBound}{Cat}(object)

\S4method{getUpperBound}{Cat}(object)

\S4method{getEstimation}{Cat}(object)

\S4method{getEstimationDefault}{Cat}(object)

\S4method{getSelection}{Cat}(object)

\S4method{getZ}{Cat}(object)

\S4method{getLengthThreshold}{Cat}(object)

\S4method{getSeThreshold}{Cat}(object)

\S4method{getInfoThreshold}{Cat}(object)

\S4method{getGainThreshold}{Cat}(object)

\S4method{getLengthOverride}{Cat}(object)

\S4method{getGainOverride}{Cat}(object)
}
\arguments{
\item{object}{An object of class \code{Cat}}
}
\value{
Returns the respective slot from Cat object
}
\description{
say something here
}
\examples{
\dontrun{
## Create Cat object
data(npi)
cat <- ltmCat(npi)

##

}

}
\seealso{
\code{\link{Cat-class}}, \code{\link{setters}}
}

