% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cat-class.R
\name{getModel}
\alias{getAnswers}
\alias{getDifficulty}
\alias{getDiscrimination}
\alias{getEstimation}
\alias{getEstimationDefault}
\alias{getGainOverride}
\alias{getGainThreshold}
\alias{getGuessing}
\alias{getInfoThreshold}
\alias{getLengthOverride}
\alias{getLengthThreshold}
\alias{getLowerBound}
\alias{getModel}
\alias{getPriorName}
\alias{getPriorParams}
\alias{getSeThreshold}
\alias{getSelection}
\alias{getUpperBound}
\alias{getZ}
\alias{setPoly}
\title{Get methods for slots within \code{Cat} class objects}
\usage{
getModel(object)
}
\arguments{
\item{object}{An object of class \code{Cat} to be modified}
}
\value{
The value stored in the specified slot of the \code{Cat} object.
}
\description{
These functions access and return the values currently stored in a specified slot of an object of class \code{Cat}.
}
\note{
There are getters for every slot in \code{Cat} objects, all following the format \code{getSlot}: 
eg. \code{getDiscrimination}, \code{getEstimationDefault}, \code{getSeThreshold}, etc.
}
\author{
Josh W. Cutler: \email{josh@zistle.com} and Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
\seealso{
{\link{Cat}} {\link{setters}}
}

