% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cat-class.R
\name{getGuessing}
\alias{getGuessing}
\title{Getters: access slots within Cat class objects}
\usage{
getGuessing(object = "Cat")
}
\arguments{
\item{object}{An object of class \code{Cat} with a slot to be changed}
}
\value{
A value of an object of class \code{Cat} with an updated slot
}
\description{
These functions allow us to access to the values currently stored in a given slot of an object of class \code{Cat} with the user-provided values.
}
\note{
Notes
}
\author{
Josh W. Cutler: \email{josh@zistle.com} and Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
\seealso{
\code{\link{Cat}}
}

