% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grmCat.R
\name{grmCat}
\alias{grmCat}
\title{Computerized Adaptive Testing Graded Response Model}
\usage{
grmCat(data, object = NULL, ...)
}
\arguments{
\item{data}{a \code{data.frame} or a numeric \code{matrix} of manifest variables.}

\item{object}{an object of class \code{Cat} to be populated. If omitted, a new object of class \code{Cat} is created.}

\item{...}{arguments to be passed to methods. For more details about the arguments, see \code{\link{grm}}.

 @return An object of class \code{Cat} with components,
\itemize{
\item \code{difficulty} a named list of difficulty parameters for use with polytomous questions/items.  Each element's name tells the question/item to which it applies.
\item \code{guessing} a vector of guessing parameter for each question/item. 
\item \code{discrimination} a vector of disrimination parameter for each question/item.
\item \code{answers} a vector of answers to questions as given by the survey respondent.
\item \code{priorName} a character vector of length one giving the prior distribution to use for the latent trait estimates.  The options are \code{normal} for the normal distirbution, \code{cauchy} for the Cauchy distribution, are \code{t} for the t-distribution. Defaults to \code{normal}. 
\item \code{priorParams} a numeric vector of parameters for the distribution specified in the \code{priorName} slot. See the details section for more infomration.  Defaults to \code{c(0,1)}.   
}}
}
\description{
This function fits the Graded Response model for ordinal polytomous data and populates the fitted values for discimination and difficulty parameters to an object of class \code{Cat}.
}
\note{
In case the Hessian matrix at convergence is not positive definite try to use \code{start.val="random"}.
}
\author{
Josh W. Cutler: \email{josh@zistle.com} and Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
\seealso{
\code{\link{ltmCat}},\code{\link{nextItem}}, \code{\link{question.path}}
}

