% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccPlots.R
\name{icc}
\alias{icc}
\title{Plotting Item Characteristic Curves (ICC)}
\usage{
icc(object, theta_range, question, ...)
}
\arguments{
\item{object}{A \code{Cat} object}

\item{theta_range}{A numerical vector of thetas at which the probabilities are to be calculated}
}
\description{
This function plots the item characteristic curves for an object of class \code{Cat}.
}

