% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood}
\alias{likelihood}
\title{Likelihood of the Specified Response Set}
\usage{
likelihood(catObj, theta)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}

\item{theta}{A numeric or an integer indicating the value for \eqn{\theta_j}}
}
\value{
The function returns a numeric value of the likelihood of the respondent having offered the provided response profile.
}
\description{
Calculates the likelihood of a respondent, with ability parameter \eqn{\theta}, having offered the specific set of responses stored in the \code{Cat} objects \code{answers} slot, conditional on the item-level parameters.
}
\details{

}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled C++ code.
}
\examples{
\dontrun{
## Likelihood for Cat object of the ltm model
data(npi)
cat <- ltmCat(npi)
setAnswers(cat) <- c(1,0,1,0, rep(NA, 35))
likelihood(cat, theta = 1)
}

 
}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\references{

}
\seealso{
\code{\link{probability}} for probability of responses to individual question item
}

