% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood}
\alias{likelihood}
\title{Likelihood of offering specific response}
\usage{
likelihood(cat_df, t)
}
\arguments{
\item{cat_df}{An object of \code{Cat} class}

\item{t}{A numeric for possible value of theta (position on the latent scale of interest)}
}
\value{
A value of the likelihood of each respondent having offered the provided response profile
}
\description{
This function returns the value of likelihood of a person with ability parameter \eqn{\theta} having offered the specific response profile stored in answers conditional on the item-level parameters.
}
\details{
Letting \deqn{q_i(\theta_j)=1-p_i(\theta_[j])} <sup>superscript</sup>
}

