% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lookAhead}
\alias{lookAhead}
\title{Look Ahead to Select Next Item}
\usage{
lookAhead(catObj, item)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}

\item{item}{A numeric indicating the question item the respondent is currently answering.}
}
\value{
A function \code{lookAhead} returns a \code{list} of one element, \code{estimates}, a \code{data.frame}.
The the first column of the \code{data.frame} is the possible response option to the question the respondent
is currently answering and the second column is the next item that should be asked given each response.
}
\description{
Selects the next item that would be asked for all possible response options to the question the respondent is currently answering.
}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled \code{C++} code.
}
\examples{
\dontrun{
## Create Cat object
data(npi)
ltm_cat <- ltmCat(npi)

## Store example answers
setAnswers(ltm_cat) <- c(1,0,1,0,1, rep(NA, 35))

## What should be asked next if respondent is currently answering item 6
lookAhead(ltm_cat, 6)
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{selectItem}}
}

