% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ltmCAT.R
\name{ltmCAT}
\alias{ltmCAT}
\title{Computerized Adaptive Testing Latent Trait Model with Binary Data}
\usage{
ltmCAT(data, object = NULL, ...)
}
\arguments{
\item{data}{a \code{data.frame} or a numeric \code{matrix} of manifest variables.}

\item{object}{an object of class \code{CATsurv} to be populated. If omitted, a new object of class \code{CATsurv} is created.}

\item{...}{arguments to be passed to methods. For more details about the arguments, see \code{\link{tpm}}.}
}
\value{
An object of class \code{CATsurv} with components,
\itemize{
\item \code{difficulty} a named vector of difficulty parameters for use with dichotomous questions/items.  Each element's name tells the question/item to which it applies.
\item \code{guessing} a vector of guessing parameter for each item.
\item \code{discrimination} a vector of disrimination parameter for each item.
\item \code{answers} a vector of answers to questions as given by the survey respondent.
\item \code{priorName} a character vector of length one giving the prior distribution to use for the latent trait estimates.  The options are \code{normal} for the normal distirbution, \code{cauchy} for the Cauchy distribution, are \code{t} for the t-distribution. Defaults to \code{normal}.
\item \code{priorParams} a numeric vector of parameters for the distribution specified in the \code{priorName} slot. See the details section for more infomration.  Defaults to \code{c(0,1)}.
}
}
\description{
This function fits the latent trait model for binary data and populates the fitted values for discimination, difficulty, and guessing parameters to an object of class \code{CATsurv}.
}
\note{
In case the Hessian matrix at convergence is not positive definite try to use \code{start.val="random"}.
}
\author{
Josh W. Cutler: \email{josh@zistle.com} and Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
\seealso{
\code{\link{grmCAT}}, \code{\link{nextItem}}, \code{\link{question.path}}
}

