% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTree.R
\name{makeTree}
\alias{makeTree}
\title{Make tree of possible question combinations}
\usage{
makeTree(cat, flat = FALSE)
}
\arguments{
\item{cat}{an object of class \code{Cat}}

\item{flat}{a logical; TRUE returns tree as a table; FALSE returns tree as a list}
}
\value{
An object of class \code{list}
}
\description{
Creates a tree of possible questions asked and stores it as a list of lists (flat = FALSE) or a table of values (flat = TRUE).
}
\details{
Function takes a \code{Cat} object and generate a tree of all possible question combinations, condtional on previous answers and current \eqn{\theta} estimates.
The tree is stored as a list of lists, iteratively generated by filling in a possible answer, calculating the next question via \code{selectItem}, filling in a possible answer for that question, and so forth.
The length of the tree is dictated by the \code{lengthThreshold} slot within the \code{Cat} object.
}
\examples{


}
\author{
Ryden W. Butler: \email{r.butler@wustl.edu} and Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
\seealso{
\code{\link{selectItem}} for specific item selection and selection methods
}

