% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTree.R
\name{makeTree}
\alias{makeTree}
\title{Make Tree of Possible Questions}
\usage{
makeTree(cat, flat = FALSE)
}
\arguments{
\item{cat}{a \code{Cat} object

 @return An object of class \code{list}}
}
\description{
This function creates a tree of possible questions asked and stores it as a list of lists.
}
\author{
Ryden W. Butler: \email{r.butler@wustl.edu} and Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
\seealso{
\code{\link{FlattenTree}},
}

