% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nfc}
\alias{nfc}
\title{Need For Cognition}
\format{A data frame with 4043 observations on the following 18 variables.
\describe{
    \item{\code{NFC1}}{I really enjoy a task that involves coming up with new solutions to problems}
    \item{\code{NFC4}}{I would prefer a task that is intellectual, difficult, and important to one that is somewhat important but does not require much thought}
    \item{\code{NFC10}}{Learning new ways to think doesn't excite me very much}
    \item{\code{NFC12}}{I usually end up deliberating about issues even when they do not affect me personally}
    \item{\code{NFC15}}{The idea of relying on thought to make my way to the top appeals to me}
    \item{\code{NFC16}}{The notion of thinking abstractly is appealing to me}
    \item{\code{NFC19}}{I only think as hard as I have to}
    \item{\code{NFC21}}{I think tasks that require little thought once I've learned them}
    \item{\code{NFC22}}{I prefer to think about small, daily projects to long-term ones}
    \item{\code{NFC23}}{I would rather do something that requires little thought than something that is sure to challenge my thinking abilities}
    \item{\code{NFC24}}{I find satisfaction in deliberating hard and for long hours}
    \item{\code{NFC29}}{I like to have the responsibility of handing a situation that requires a lot of thinkings}
    \item{\code{NFC31}}{I feel relief rather than satisfaction after completing a task that required a lot of mental effort}
    \item{\code{NFC32}}{Thinking is not my idea of fun}
    \item{\code{NFC33}}{I try to anticipate and avoid situations where there is likely a chance I will have to think in depth about something}
    \item{\code{NFC39}}{I prefer my life to be filed with puzzles that I must solve}
    \item{\code{NFC40}}{I would prefer complex to simple problems}
    \item{\code{NFC43}}{Its enough for me that something gets the job done; I don't care how or why it works}
}}
\usage{
data(nfc)
}
\description{
Data of survey respondents' responses to 18 NFC questions, which is a reduced version of NFC (Cacioppo and Petty 1984). For each question, respondents could choose one of five response options: 1 = agree strongly, 2 = agree somewhat, 3 = neither agree nor disagree, 4 = disagree somewhat, 5 = disagree strongly. Missingness is optional.
Observations come from one of three administrations of the survey: The American Panel Survey, Amazon's Mechanical Turk  in fall of 2014, or Amazon's Mechanical Turk in spring of 2015.
}
\references{
Cacioppo, John T. and Richard E. Petty. 1984. "The Efficient Assessment of Need for Cognition." Journal of Personality Assessment 48(3):306-307.
}
\keyword{datasets}
