% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{npi_battery}
\alias{npi_battery}
\title{Narcissistic personality inventory question items}
\format{
An object of class \code{list} of length 40.
}
\usage{
data(npi_battery)
}
\description{
A list where each element is a length three vector containin the text of the the question item and the two response options.
}
\examples{

data(npi_battery)

# Item 1
npi_battery$Q1[1]

# Response options for item 1
npi_battery$Q1[2:3]


}
\references{
Raskin, R., and H. Terry (1988), "A Principal-Components Analysis of the Narcissistic Personality
Inventory and Further Evidence of Its Construct Validity," Journal of Personality and Social
Psychology, 54, 890-902.
}
\seealso{
\code{\link{npi_cat}}
}
\keyword{datasets}
