% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obsInf}
\alias{obsInf}
\title{Observed Information}
\usage{
obsInf(catObj, theta, item)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}

\item{theta}{A numeric or an integer indicating the value for \eqn{\theta_j}}

\item{item}{An integer indicating the index of the question item}
}
\value{
The function returns a numeric value of the observed information of the likelihood, given \eqn{\theta}, for a specific question item
}
\description{
Calculates the observed information of the likelihood evaluated at the input value \eqn{\theta} for a specific item
}
\details{
The observed information is equivalent to the negative second derivative of the log-likelihood.
  Note: This function should never be called when the respondent has answered no questions.
}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled C++ code.
}
\examples{
\dontrun{
## Prior calculation using Cat object of the ltm model
## specifying different distributions
data(npi)
cat <- ltmCat(npi)
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{estimateTheta}} for calculation of \eqn{\theta} and
  \code{\link{expectedObsInf}} for further application of observed information
}

