% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obsInf}
\alias{obsInf}
\title{Observed Information}
\usage{
obsInf(catObj, theta, item)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}

\item{theta}{A numeric or an integer indicating the value for \eqn{\theta}}

\item{item}{An integer indicating the index of the question item}
}
\value{
The function \code{obsInf} returns a numeric value of the observed information of the likelihood, given \eqn{\theta}, for a given question item.
}
\description{
Calculates the observed information of the likelihood of a respondent's ability \eqn{\theta} for a given \code{item}.
}
\details{
The observed information is equivalent to the negative second derivative of the log-likelihood evaluated at \eqn{\theta}.
This function should never be called when the respondent has answered no questions as the likelihood is not defined.
}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled \code{C++} code.
}
\examples{
\dontrun{
## Create Cat object, store example answers, and calculate observed information
for an ability parameter of 1 for item 10

## observed information for Cat object of the ltm model
data(npi)
ltm_cat <- ltmCat(npi)
setAnswers(ltm_cat) <- c(1,0,1,0,1, rep(NA, 35))
obsInf(ltm_cat, theta = 1, item = 10)

## observed information for Cat object of the tpm model
data(polknow)
tpm_cat <- tpmCat(polknow)
setAnswers(tpm_cat) <- c(1,0,1,0, rep(NA, 35))
obsInf(tpm_cat, theta = 1, item = 10)

## observed information for Cat object of the grm model
data(nfc)
grm_cat <- grmCat(nfc)
setAnswers(grm_cat) <- c(1,3,4,5, rep(NA, 13))
obsInf(grm_cat, theta = 1, item = 10)
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{estimateTheta}}, \code{\link{expectedObsInf}}
}

