% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Cat}
\alias{plot.Cat}
\alias{plot}
\title{Plotting function for Cat object}
\usage{
plot.Cat(x, item, type, ...)
}
\arguments{
\item{x}{Cat.}

\item{item}{Numeric.}

\item{type}{Character.}

\item{...}{Other arguments passed to plot().}
}
\description{
Description.
}
