% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Cat.R
\name{plot.Cat}
\alias{plot.Cat}
\alias{plot,Cat}
\alias{plot,Cat-method}
\title{Plotting function for Cat object}
\usage{
\S4method{plot}{Cat}(x, item, plotType, xlim = c(-5, 5), ...)
}
\arguments{
\item{x}{Cat object.}

\item{item}{Numeric. Item index.}

\item{plotType}{Character. Either "IRF", "ICC", or "IIF" for item response functions, item characteristic functions, and item information functions, respectively.}

\item{xlim}{Vector. Range of the x-axis.}

\item{...}{Other arguments passed to plot().}
}
\description{
Function to plot item response functions, item characteristic functions, and item information functions
for Cat objects.
}
