% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Cat.R
\name{plot.Cat}
\alias{plot.Cat}
\alias{plot,Cat}
\alias{plot,Cat-method}
\title{Plotting function for Cat object}
\usage{
\S4method{plot}{Cat}(x, item, plotType, xlim = c(-5, 5), ...)
}
\arguments{
\item{x}{Cat.}

\item{item}{Numeric.}

\item{plotType}{Character.}

\item{xlim}{Vector.}

\item{...}{Other arguments passed to plot().}
}
\description{
Description.
}
