% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posteriorKL}
\alias{posteriorKL}
\title{Expected Kullback-Leibeler information, weighted by the posterior}
\usage{
posteriorKL(catObj, item)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}

\item{item}{An integer indicating the index of the question item}
}
\value{
A value indicating the posterior KL information for the desired item, given the current answer profile and ability estimate.
}
\description{
Calculate the expected Kullback-Leibeler information, weighted by the posterior
}
\details{
We will follow the same procedure as \code{expectedKL}, except we will weight the different potential values of \eqn{\theta_0} by the posterior.

Due to the conditional independence assumption, we only need to calculate the expected value for potential new items.
}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled C++ code.
}
\examples{
\dontrun{
## Create Cat object
data(npi)
ltm_cat <- ltmCat(npi)

## Store example answers
setAnswers(ltm_cat) <- c(1,0,1,0,1, rep(NA, 35))
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{likelihoodKL}} and/or \code{\link{expectedKL}} for alternative KL methods
}

