% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prior}
\alias{prior}
\title{Prior Value}
\usage{
prior(x, dist, params)
}
\arguments{
\item{x}{A numeric value at which to evaluate the prior}

\item{dist}{A string indicating the distribution (slot \code{priorName} of \code{Cat} object)}

\item{params}{A length two numeric vector indicating the parameters of the distribution (slot \code{priorParams} of \code{Cat} object)}
}
\value{
A scalar consisting of prior value, \eqn{\pi(x)}, given the value \eqn{x}
}
\description{
Calculates the density at \code{x} of either the normal, student's t, or uniform distribution
}
\details{
The \eqn{dist} argument needs to be either "UNIFORM", "NORMAL", or "STUDENT_T".

When \eqn{dist} is "NORMAL", the first element of \eqn{params} is the mean, 
the second element is the standard deviation.  When \eqn{dist} is "STUDENT_T", the first 
element of \eqn{params} is the non-centrality parameters and the second is degrees of freedom.  
When \eqn{dist} is "UNIFORM", the elements of \eqn{params} are the lower and upper bounds,
of the interval, respectively.  Note that the "UNIFORM" is only applicable for the "EAP" estimation method.
}
\note{
This function is to allow users to access the internal functions of the package. During item selection, all calculations are done in compiled C++ code.

This function uses Boost C++ source libraries for the uniform and Student's t
distributions and calls \code{dnorm4} written in C which is identical to that 
of \code{dnorm} in \code{R}.
}
\examples{
\dontrun{
## Prior calculation using Cat object of the ltm model
## specifying different distributions
data(npi)
cat <- ltmCat(npi)

cat@priorName <- "NORMAL"
cat@priorParams <- c(0, 1) ## Parameters are mean and standard deviation
prior(x = 1, cat@priorName, cat@priorParams)

cat@priorName <- "STUDENT_T"
cat@priorParams <- c(1, 3) ## Parameters are non-centrality param and degrees of freedom
prior(x = 1, cat@priorName, cat@priorParams)

cat@priorName <- "UNIFORM"
cat@priorParams <- c(-1, 1) ## Parameters are lower bound and upper bound of interval
prior(x = 1, cat@priorName, cat@priorParams)
}

}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery,
 Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
\seealso{
\code{\link{Cat}} for information on \code{priorName} and \code{priorParams} slots
}

