% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prior}
\alias{prior}
\title{The possible prior distribution functions}
\usage{
prior(x, c, p)
}
\arguments{
\item{x}{A numeric value where we want to evaluate the prior name}
}
\value{
A vector consisting of prior value, \eqn{\pi(x)}, given the value \eqn{x}
}
\description{
This function returns the prior value for each respondent's position on the latent scale of interest
}
\details{
\eqn{x} needs to be either NORMAL or STUDENT_T params, which are parameters controlling the shape of the prior
}

