% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prior}
\alias{prior}
\title{The possible prior distribution functions}
\usage{
prior(x, c, p)
}
\arguments{
\item{x}{A numeric value where we want to evaluate the prior name}

\item{c}{The @priorName slot of a \code{Cat} object}

\item{p}{The @priorParams slot of a \code{Cat} object}
}
\value{
A vector consisting of prior value, \eqn{\pi(x)}, given the value \eqn{x}
}
\description{
This function returns the prior value for each respondent's position on the latent scale of interest
}
\details{
Note: \eqn{x} needs to be either UNIFORM, NORMAL, or STUDENT_T parameters, which control the shape of the prior.
}

