% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{probability}
\alias{probability}
\title{Probabilities of the responses to a question given theta}
\usage{
probability(cat_df, t, q)
}
\arguments{
\item{cat_df}{An object of \code{Cat} class}

\item{t}{A double indicating the potential value for \eqn{\theta_j}}

\item{q}{An integer indicating the index of the question}

\item{ret_prob}{(For polytonomous implementation only) A double-vector where the calculations carried out by this function will be stored.}
}
\value{
A vector consisting of the probability of a correct response for each respondent on item \eqn{i}.
}
\description{
This function calculates the probabilities of a specific set of responses to a specific question for a specific value of \eqn{\theta}.
}
\details{
The probability of a correct response for respondent \eqn{j} on item \eqn{i} is ....
where \eqn{\theta_j} is respondent \eqn{j}'s position on the latent scale of interest, \eqn{a_i} is item \eqn{i}'s discrimination parameter,
 \eqn{b_i} is item i's difficulty parameter, and \eqn{c_i} is item \eqn{i}'s guessing parameter.

 Note: this function is overloaded, due to different output types of binary vs polytomous implementations (outputs single value for binary implementation,
 vector of values for polytomous implementation)

 Note: the function for polytomous implementation does not return values, but rather alters the object ret_prob in memory
}

