% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{selectItem}
\alias{selectItem}
\title{Select the next item in the question set}
\usage{
selectItem(catObj)
}
\arguments{
\item{catObj}{An object of class \code{Cat}}
}
\value{
It returns a list with two elements: 
(1) A dataframe containing a column with the indexes of unasked questions and a column with the values (calculated by the specified selection method) for those items, 
and (2) a numeric containing the index of the question that should be asked next.
}
\description{
Select the next item in the question set based on the specified method
}
\details{
This function takes in a \code{Cat} object from \code{R} and constructs the \code{C++} representation.

  The EPV method:
  
  This function takes in a \code{Cat} object from \code{R} and constructs the \code{C++} representation. It then calculates the expected posterior variance for each unanswered item. 
  
  The function returns a list with the following two elements:
  \code{all.estimates}: A data fame with two columns. The first column ("questions") should be the index of the question items and the second column ("EPV") of the expected posterior variance for that corresponding item. 
  There are as many rows in this data frame as there are unsanswered questions in the battery.
  \code{next.item}: A numeric vector with the index of the item with the lowest EPV value.
  
  See \code{expectedPV} for mathematical details.
  
  The MFI method:
  
  This function takes a \code{Cat} object and calculates Fisher's information for each unanswered item. It then finds the one item that maximizes Fisher's information, based on the respondent's position on the latent trait from the answered items. 
  
  The function returns a list with the following two elements:
  \code{all.estimates}: A data fame with two columns. The first column ("questions") should be the index of the question items and the second column of the expected posterior variance for that corresponding item.
  There are as many rows in this data frame as there are unsanswered questions in the battery.
  \code{next.item}: A numeric vector with the index of the item with the highest Fisher's information value.

  See \code{fisherInf} for mathematical details.
  
  The MLWI method:
  
  This function calculates the likelihood for each value of X at the input value of \eqn{\theta}.
  Evaluates the integral over a measure of the plausibility of possible values of \eqn{\theta} by weighting Fisher's information with the likelihood function and selecting the next question according to:
 
 \deqn{i_j = \underset{j}{\operatorname{argmax}}\int_{-\infty}^{\infty} L(\theta_j)I_i(\theta_j)d\theta_j}{i_j = argmax_j \int_{-\infty}^{\infty} L(\theta_j) I_i(\theta_j) d\theta_j}
  
  The MPWI method:
  
  This function calculates the likelihood for each value of X at the input value of \eqn{\theta}.
  Evaluates the integral over a measure of the plausibility of possible values of \eqn{\theta} by weighting Fisher's information with the likelihood function and selecting the next question according to:

  \deqn{i_j = \underset{j}{\operatorname{argmax}}\int_{-\infty}^{\infty} \pi(\theta) L(\theta_j)I_i(\theta_j)d\theta_j}{i_j = argmax_j \int_{-\infty}^{\infty} \pi(\theta) L(\theta_j) I_i(\theta_j) d\theta_j}

  The MEI method:
  
  This function estimates the expected observed information for a respondent’s estimated
  position on the latent trait on an unanswered item based on the
  respondent’s position on the latent trait calculated from answered items.
  
  The output should be a single numeric value.

  Binary details:

  \deqn{i_j = \underset{j}{\operatorname{argmax}}\{P(y_{ij} = 1)J_{y_{ij}=1} + P(y_{ij} = 0)J_{y_{ij}=0} \}}{i_j = argmax_j {P(y_ij = 1) J_{y_ij = 1} + P(y_ij = 0) J_(y_ij = 0)}}

  Categorical details:
  
  \deqn{i_j = \underset{j}{\operatorname{argmax}}\{\sum_k^K P(y_{ij} = k)J_{y_{ij}=k} \}}{i_j = argmax_j {\sum_k^K P(y_ij = k) J_{y_ij = k}}}
  
  The KL method:
  
  This procedure chooses the next item with the largest KL value.
  
  See \code{expectedKL}, \code{likelihoodKL}, and/or \code{posteriorKL} for mathematical details.
  
  The MFII method:
  
  This approach chooses items based on the Fisher's information in an interval near the current estimate \eqn{\hat{\theta}}.
  
  \deqn{FII_i = \int^{\hat{\theta}+\delta}_{\hat{\theta}-\delta}I_i(\theta_0) d\theta_0}{FII_i = \int^{\hat{\theta} + \delta}_{\hat{\theta} - \delta} I_i(\theta_0) d\theta_0}

  where \deqn{\delta = z(I(\hat{\theta}))^{-1/2}}{\delta = z(I(\hat{\theta}))^{-1/2}},  \eqn{I(\hat{\theta})} is the test information for respondent \eqn{j} evaluated at \eqn{\hat{\theta}},  \deqn{I_i(\cdot)}{I_i(.)} is the Fisher's information for item \eqn{i}, and \eqn{z} is a user specified z-value.
  
  The random method:
  
  This routine serves as a baseline for comparison. The routine simply selects an unanswered question at random.
}
\examples{

 
}
\seealso{
\code{\link{estimateTheta}} for calculation of \eqn{\theta};  
  \code{\link{obsInf}} for observed information calculation;
  \code{\link{fisherTestInfo}} for Fisher's information calculation;
  \code{\link{expctedKL}} for expected Kullback-Leibeler calculation;
  \code{\link{likelihoodKL}} for likelihood Kullback-Leibeler calculation;  
  \code{\link{posteriorKL}} for posterior Kullback-Leibeler calculation;
}

