% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{selectItem}
\alias{selectItem}
\title{Select the next item in the question set}
\usage{
selectItem(cat_df)
}
\arguments{
\item{cat_df}{An object of \code{Cat} class}
}
\value{
It returns a list with two elements: 
(1) A dataframe containing a column with the indexes of unasked questions and a column with the values (calculated by the specified selection method) for those items, 
and (2) a numeric containing the index of the question that should be asked next.
}
\description{
Select the next item in the question set based on the specified method
}
\details{
This function takes in a \code{Cat} object from \code{R} and constructs the \code{C++} representation.

  For the EPV method:
  
  This function takes in a \code{Cat} object from \code{R} and constructs the \code{C++} representation. It then calculates the expected posterior variance for each unanswered item. 
  
  The function returns a list with the following two elements:
  \code{all.estimates}: A data fame with two columns. The first column ("questions") should be the index of the question items and the second column ("EPV") of the expected posterior variance for that corresponding item. 
  There are as many rows in this data frame as there are unsanswered questions in the battery.
  \code{next.item}: A numeric vector with the index of the item with the lowest EPV value.
  
  See \code{expectedPV} for mathematical details.
  
  For the MFI method:
  
  This function takes a \code{Cat} object and calculates Fisher's information for each unanswered item. It then finds the one item that maximizes Fisher's information, based on the respondent's position on the latent trait from the answered items. 
  
  The function returns a list with the following two elements:
  \code{all.estimates}: A data fame with two columns. The first column ("questions") should be the index of the question items and the second column of the expected posterior variance for that corresponding item.
  There are as many rows in this data frame as there are unsanswered questions in the battery.
  \code{next.item}: A numeric vector with the index of the item with the highest Fisher's information value.

  See \code{fisherInf} for mathematical details.
  
  For the MLWI method:
  
  For the MPWI method:
  
  For the MEI method:
  
  For the KL method:
  
  For the MFII method:
  
  For the random method:
}

