% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cat-class.R
\docType{methods}
\name{setters}
\alias{setAnswers<-}
\alias{setAnswers<-,Cat-method}
\alias{setDifficulty<-}
\alias{setDifficulty<-,Cat-method}
\alias{setDiscrimination<-}
\alias{setDiscrimination<-,Cat-method}
\alias{setEstimation<-}
\alias{setEstimation<-,Cat-method}
\alias{setEstimationDefault<-}
\alias{setEstimationDefault<-,Cat-method}
\alias{setGainOverride<-}
\alias{setGainOverride<-,Cat-method}
\alias{setGainThreshold<-}
\alias{setGainThreshold<-,Cat-method}
\alias{setGuessing<-}
\alias{setGuessing<-,Cat-method}
\alias{setInfoThreshold<-}
\alias{setInfoThreshold<-,Cat-method}
\alias{setLengthOverride<-}
\alias{setLengthOverride<-,Cat-method}
\alias{setLengthThreshold<-}
\alias{setLengthThreshold<-,Cat-method}
\alias{setLowerBound<-}
\alias{setLowerBound<-,Cat-method}
\alias{setModel<-}
\alias{setModel<-,Cat-method}
\alias{setPriorName<-}
\alias{setPriorName<-,Cat-method}
\alias{setPriorParams<-}
\alias{setPriorParams<-,Cat-method}
\alias{setSeThreshold<-}
\alias{setSeThreshold<-,Cat-method}
\alias{setSelection<-}
\alias{setSelection<-,Cat-method}
\alias{setUpperBound<-}
\alias{setUpperBound<-,Cat-method}
\alias{setZ<-}
\alias{setZ<-,Cat-method}
\alias{setters}
\title{Methods for Setting Value(s) to Slot in Cat Object}
\usage{
\S4method{setGuessing}{Cat}(object) <- value

\S4method{setDiscrimination}{Cat}(object) <- value

\S4method{setDifficulty}{Cat}(object) <- value

\S4method{setAnswers}{Cat}(object) <- value

\S4method{setModel}{Cat}(object) <- value

\S4method{setPriorName}{Cat}(object) <- value

\S4method{setPriorParams}{Cat}(object) <- value

\S4method{setLowerBound}{Cat}(object) <- value

\S4method{setUpperBound}{Cat}(object) <- value

\S4method{setEstimation}{Cat}(object) <- value

\S4method{setEstimationDefault}{Cat}(object) <- value

\S4method{setSelection}{Cat}(object) <- value

\S4method{setZ}{Cat}(object) <- value

\S4method{setLengthThreshold}{Cat}(object) <- value

\S4method{setSeThreshold}{Cat}(object) <- value

\S4method{setGainThreshold}{Cat}(object) <- value

\S4method{setInfoThreshold}{Cat}(object) <- value

\S4method{setLengthOverride}{Cat}(object) <- value

\S4method{setGainOverride}{Cat}(object) <- value
}
\arguments{
\item{object}{An object of class \code{Cat}}

\item{value}{The value(s) to fill the slot}
}
\value{
Updates the respective slot in Cat object
}
\description{
say something here
}
\examples{
\dontrun{
## Create Cat object
data(npi)
cat <- ltmCat(npi)

##

}

}
\seealso{
\code{\link{Cat-class}}, \code{\link{getters}}
}

