% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cat-class.R
\name{setGuessing<-}
\alias{setAnswers}
\alias{setDifficulty}
\alias{setDiscrimination}
\alias{setEstimation}
\alias{setEstimationDefault}
\alias{setGainOverride}
\alias{setGainThreshold}
\alias{setGuessing}
\alias{setGuessing<-}
\alias{setInfoThreshold}
\alias{setLengthOverride}
\alias{setLengthThreshold}
\alias{setLowerBound}
\alias{setPoly}
\alias{setPriorName}
\alias{setPriorParams}
\alias{setSeThreshold}
\alias{setSelection}
\alias{setUpperBound}
\alias{setZ}
\title{Set methods for slots within \code{Cat} class objects}
\usage{
setGuessing(object) <- value
}
\arguments{
\item{object}{An object of class \code{Cat} to be modified}

\item{value}{The value to replace the current value stored in the specified slot of the \code{Cat}}

\item{valid}{Boolean for whether to check validity of object before allowing setter method to execute, default to TRUE}
}
\value{
Modifies the \code{Cat} object that was provided by the user, and returns the modified object
}
\description{
This functions replaces the values currently stored in a specified slot of an object of class \code{Cat} with the user-provided values.
}
\note{
There are setters for every slot in \code{Cat} objects, all following the format \code{setSlot}: 
eg. \code{setDiscrimination}, \code{setEstimationDefault}, \code{setSeThreshold}, etc.
}
\author{
AA
}
\seealso{
\code{\link{Cat}}
}

