% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateRespondents.R
\docType{methods}
\name{simulateRespondents}
\alias{simulateRespondents}
\alias{simulateRespondents,Cat-method}
\title{Simulate Answer Profiles}
\usage{
\S4method{simulateRespondents}{Cat}(cat, theta, n)
}
\arguments{
\item{theta}{A numeric representing the true position on the latent trait.}

\item{n}{A numeric indicating the number of answer profiles to simulate.}

\item{catObj}{An object of class \code{Cat}}
}
\value{
...
}
\description{
Generating n answer profiles given a true value of theta and a battery's item parameters.
}
\details{
....
}
\author{
Haley Acevedo, Ryden Butler, Josh W. Cutler, Matt Malis, Jacob M. Montgomery, Tom Wilkinson, Erin Rossiter, Min Hee Seo, Alex Weil
}
