% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storeAnswer.R
\name{storeAnswer}
\alias{storeAnswer}
\title{Computerized Adaptive Testing Survey Store Answer Function}
\usage{
storeAnswer(cat, item, answer)
}
\arguments{
\item{cat}{An object of class \code{Cat}}

\item{item}{The question for which to estimate the expected posterior variance for a respondent with a latent trait estimate of theta.hat.  This should be the name of a row in the "questions" data-frame in the "questions" slot of a \code{Cat} object.}

\item{answer}{The answer to the item \emph{k} to be stored}
}
\value{
An updated object of class \code{Cat} containing the answers to \emph{k} items
}
\description{
This function updates the object of class \code{Cat} by storing the answer to item \emph{k} into questions data frame.
}
\author{
Josh W. Cutler: \email{josh@zistle.com} and Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
\seealso{
\code{\link{three.pl}},\code{\link{likelihood}}, \code{\link{prior.value}}, \code{\link{estimateTheta}}, \code{\link{estimateSE}}, \code{\link{expectedPV}},  \code{\link{nextItem}}, \code{\link{debugNextItem}}
}

